Cardano.Ledger.Api.Governance

Documentation

class (EraPParams era, Eq (GovState era), Show (GovState era), NoThunks (GovState era), NFData (GovState era), EncCBOR (GovState era), DecCBOR (GovState era), DecShareCBOR (GovState era), ToCBOR (GovState era), FromCBOR (GovState era), Default (GovState era), ToJSON (GovState era)) ⇒ EraGov era

Minimal complete definition

curPParamsGovStateL, prevPParamsGovStateL, obligationGovState

Associated Types

type GovState era = (r ∷ Type) | r → era

emptyGovState ∷ EraGov era ⇒ GovState era

Construct empty governance state

getProposedPPUpdates ∷ EraGov era ⇒ GovState era → Maybe (ProposedPPUpdates era)

Returns Nothing for all eras starting with Conway, otherwise returns proposed pparams updates

Shelley

data ShelleyGovState era

Constructors

ShelleyGovState

Fields

sgsCurProposals ∷ !(ProposedPPUpdates era)

sgsFutureProposals ∷ !(ProposedPPUpdates era)

sgsCurPParams ∷ !(PParams era)

sgsPrevPParams ∷ !(PParams era)

newtype ProposedPPUpdates era

Update operation for protocol parameters structure PParams

Constructors

ProposedPPUpdates (Map (KeyHash 'Genesis (EraCrypto era)) (PParamsUpdate era))

emptyPPPUpdates ∷ ProposedPPUpdates era

Conway

Governance Procedures

data VotingProcedure era

Constructors

VotingProcedure

Fields

vProcVote ∷ !Vote

vProcAnchor ∷ !(StrictMaybe (Anchor (EraCrypto era)))

newtype VotingProcedures era

Constructors

VotingProcedures

Fields

unVotingProcedures ∷ Map (Voter (EraCrypto era)) (Map (GovActionId (EraCrypto era)) (VotingProcedure era))

data ProposalProcedure era

Constructors

ProposalProcedure

Fields

pProcDeposit ∷ !Coin

pProcReturnAddr ∷ !(RewardAccount (EraCrypto era))

pProcGovAction ∷ !(GovAction era)

pProcAnchor ∷ !(Anchor (EraCrypto era))

Constitution

data Constitution era

Constructors

Constitution

Fields

constitutionAnchor ∷ !(Anchor (EraCrypto era))

constitutionScript ∷ !(StrictMaybe (ScriptHash (EraCrypto era)))

constitutionAnchorL ∷ Lens' (Constitution era) (Anchor (EraCrypto era))

constitutionScriptL ∷ Lens' (Constitution era) (StrictMaybe (ScriptHash (EraCrypto era)))

Governance State

data ConwayGovState era

Constructors

ConwayGovState

Fields

cgsProposals ∷ !(Proposals era)

cgsCommittee ∷ !(StrictMaybe (Committee era))

cgsConstitution ∷ !(Constitution era)

cgsCurPParams ∷ !(PParams era)

cgsPrevPParams ∷ !(PParams era)

cgsDRepPulsingState ∷ !(DRepPulsingState era)

The cgsDRepPulsingState field is a pulser that incrementally computes the stake distribution of the DReps over the Epoch following the close of voting at end of the previous Epoch. It assembles this with some of its other internal components into a (RatifyEnv era) when it completes, and then calls the RATIFY rule and eventually returns the updated RatifyState. The pulser is created at the Epoch boundary, but does no work until it is pulsed in the NEWEPOCH rule, whenever the system is NOT at the epoch boundary.

cgsProposalsL ∷ Lens' (ConwayGovState era) (Proposals era)

data RatifyState era

Constructors

RatifyState

Fields

rsEnactState ∷ !(EnactState era)

rsEnacted ∷ !(Seq (GovActionState era))

rsExpired ∷ !(Set (GovActionId (EraCrypto era)))

rsDelayed ∷ !Bool

data EnactState era

Constructors

EnactState

Fields

ensCommittee ∷ !(StrictMaybe (Committee era))

Constitutional Committee

ensConstitution ∷ !(Constitution era)

Constitution

ensCurPParams ∷ !(PParams era)

ensPrevPParams ∷ !(PParams era)

ensTreasury ∷ !Coin

ensWithdrawals ∷ !(Map (Credential 'Staking (EraCrypto era)) Coin)

ensPrevGovActionIds ∷ !(GovRelation StrictMaybe era)

Last enacted GovAction Ids

data Voter c

Constructors

CommitteeVoter !(Credential 'HotCommitteeRole c)

DRepVoter !(Credential 'DRepRole c)

StakePoolVoter !(KeyHash 'StakePool c)

data Vote

Constructors

VoteNo

VoteYes

Abstain

Governance Action

data GovAction era

Note that the previous governance action id is only optional for the very first governance action of the same purpose.

Constructors

ParameterChange

Fields

!(StrictMaybe (GovPurposeId 'PParamUpdatePurpose era))

Previous governance action id of ParameterChange type, which corresponds to PParamUpdatePurpose.

!(PParamsUpdate era)

Proposed changes to PParams

!(StrictMaybe (ScriptHash (EraCrypto era)))

Policy hash protection

HardForkInitiation

Fields

!(StrictMaybe (GovPurposeId 'HardForkPurpose era))

Previous governance action id of HardForkInitiation type, which corresponds to HardForkPurpose

!ProtVer

Proposed new protocol version

TreasuryWithdrawals

Fields

!(Map (RewardAccount (EraCrypto era)) Coin)

Proposed treasury withdrawals

!(StrictMaybe (ScriptHash (EraCrypto era)))

Policy hash protection

NoConfidence

Fields

!(StrictMaybe (GovPurposeId 'CommitteePurpose era))

Previous governance action id of NoConfidence or UpdateCommittee type, which corresponds to CommitteePurpose

UpdateCommittee

Fields

!(StrictMaybe (GovPurposeId 'CommitteePurpose era))

Previous governance action id of UpdateCommittee or NoConfidence type, which corresponds to CommitteePurpose

!(Set (Credential 'ColdCommitteeRole (EraCrypto era)))

Constitutional Committe members to be removed

!(Map (Credential 'ColdCommitteeRole (EraCrypto era)) EpochNo)

Constitutional committee members to be added

!UnitInterval

New Threshold

NewConstitution

Fields

!(StrictMaybe (GovPurposeId 'ConstitutionPurpose era))

Previous governance action id of NewConstitution type, which corresponds to ConstitutionPurpose

!(Constitution era)

InfoAction

data GovActionId c

Constructors

GovActionId

Fields

gaidTxId ∷ !(TxId c)

gaidGovActionIx ∷ !GovActionIx

newtype GovActionIx

Constructors

GovActionIx

Fields

unGovActionIx ∷ Word32

data GovActionState era

Constructors

GovActionState

Fields

gasId ∷ !(GovActionId (EraCrypto era))

gasCommitteeVotes ∷ !(Map (Credential 'HotCommitteeRole (EraCrypto era)) Vote)

gasDRepVotes ∷ !(Map (Credential 'DRepRole (EraCrypto era)) Vote)

gasStakePoolVotes ∷ !(Map (KeyHash 'StakePool (EraCrypto era)) Vote)

gasProposalProcedure ∷ !(ProposalProcedure era)

gasProposedIn ∷ !EpochNo

gasExpiresAfter ∷ !EpochNo

data GovActionPurpose

Constructors

PParamUpdatePurpose

HardForkPurpose

CommitteePurpose

ConstitutionPurpose

data GovRelation (f ∷ Type → Type) era

Abstract data type for representing relationship of governance action with the same purpose

Constructors

GovRelation

Fields

grPParamUpdate ∷ !(f (GovPurposeId 'PParamUpdatePurpose era))

grHardFork ∷ !(f (GovPurposeId 'HardForkPurpose era))

grCommittee ∷ !(f (GovPurposeId 'CommitteePurpose era))

grConstitution ∷ !(f (GovPurposeId 'ConstitutionPurpose era))

hoistGovRelation ∷ (∀ a. f a → g a) → GovRelation f era → GovRelation g era

withGovActionParent

Arguments

∷ GovActionState era

→ a
The result to be used for governance actions that can't have a parent

→ (∀ (p ∷ GovActionPurpose). (∀ (f ∷ Type → Type). Lens' (GovRelation f era) (f (GovPurposeId p era))) → StrictMaybe (GovPurposeId p era) → GovPurposeId p era → a)
Function that will be applied to a lens and a parent

→ a

Apply a function to a GovAction that can have a parent.

newtype GovPurposeId (p ∷ GovActionPurpose) era

Constructors

GovPurposeId

Fields

unGovPurposeId ∷ GovActionId (EraCrypto era)

govActionIdToText ∷ GovActionId c → Text

Anchor

data Anchor c

Constructors

Anchor

Fields

anchorUrl ∷ !Url

anchorDataHash ∷ !(SafeHash c AnchorData)

newtype AnchorData

Constructors

AnchorData ByteString

hashAnchorData ∷ Crypto c ⇒ AnchorData → SafeHash c AnchorData

Hash AnchorData