Cardano.Ledger.Api.State.Query

GetFilteredDelegationsAndRewardAccounts

queryStakePoolDelegsAndRewards ∷ EraCertState era ⇒ NewEpochState era → Set (Credential 'Staking) → (Map (Credential 'Staking) (KeyHash 'StakePool), Map (Credential 'Staking) Coin)

Implementation for GetFilteredDelegationsAndRewardAccounts query.

GetGovState

queryGovState ∷ NewEpochState era → GovState era

This query returns all of the state related to governance

GetConstitution

queryConstitution ∷ ConwayEraGov era ⇒ NewEpochState era → Constitution era

GetConstitutionHash

queryConstitutionHash ∷ ConwayEraGov era ⇒ NewEpochState era → SafeHash AnchorData

GetDRepState

queryDRepState

Arguments

∷ ConwayEraCertState era

⇒ NewEpochState era

→ Set (Credential 'DRepRole)
Specify a set of DRep credentials whose state should be returned. When this set is empty, states for all of the DReps will be returned.

→ Map (Credential 'DRepRole) DRepState

Query DRep state.

GetDRepDelegations

queryDRepDelegations

Arguments

∷ ∀ era. ConwayEraCertState era

⇒ NewEpochState era

→ Set DRep
Specify a set of DReps whose state should be returned. When this set is empty, states for all of the DReps will be returned.

→ Map DRep (Set (Credential 'Staking))

Query the delegators delegated to each DRep, including AlwaysAbstain and NoConfidence.

GetDRepStakeDistr

queryDRepStakeDistr

Arguments

∷ ConwayEraGov era

⇒ NewEpochState era

→ Set DRep
Specify DRep Ids whose stake distribution should be returned. When this set is empty, distributions for all of the DReps will be returned.

→ Map DRep Coin

Query DRep stake distribution. Note that this can be an expensive query because there is a chance that current distribution has not been fully computed yet.

GetRegisteredDRepStakeDistr

queryRegisteredDRepStakeDistr

Arguments

∷ (ConwayEraGov era, ConwayEraCertState era)

⇒ NewEpochState era

→ Set (Credential 'DRepRole)
Specify DRep Ids whose stake distribution should be returned. When this set is empty, distributions for all of the registered DReps will be returned.

→ Map (Credential 'DRepRole) Coin

Query the stake distribution of the registered DReps. This does not include the AlwaysAbstain and NoConfidence DReps.

GetSPOStakeDistr

querySPOStakeDistr

Arguments

∷ ConwayEraGov era

⇒ NewEpochState era

→ Set (KeyHash 'StakePool)

→ Map (KeyHash 'StakePool) Coin
Specify pool key hashes whose stake distribution should be returned. When this set is empty, distributions for all of the pools will be returned.

Query pool stake distribution.

GetCommitteeState

queryCommitteeState ∷ ConwayEraCertState era ⇒ NewEpochState era → CommitteeState era

Deprecated: In favor of queryCommitteeMembersState

Query committee members

GetCommitteeMembersState

queryCommitteeMembersState

Arguments

∷ ∀ era. (ConwayEraGov era, ConwayEraCertState era)

⇒ Set (Credential 'ColdCommitteeRole)
filter by cold credentials (don't filter when empty)

→ Set (Credential 'HotCommitteeRole)
filter by hot credentials (don't filter when empty)

→ Set MemberStatus
filter by status (don't filter when empty) (useful, for discovering, for example, only active members)

→ NewEpochState era

→ CommitteeMembersState

Query committee members. Whenever the system is in No Confidence mode this query will return Nothing.

GetChainAccountState

queryChainAccountState ∷ NewEpochState era → ChainAccountState

data CommitteeMemberState

Constructors

CommitteeMemberState

Fields

cmsHotCredAuthStatus ∷ !HotCredAuthStatus

cmsStatus ∷ !MemberStatus

cmsExpiration ∷ !(Maybe EpochNo)

Absolute epoch number when the member expires

cmsNextEpochChange ∷ !NextEpochChange

Changes to the member at the next epoch

data CommitteeMembersState

Constructors

CommitteeMembersState

Fields

csCommittee ∷ !(Map (Credential 'ColdCommitteeRole) CommitteeMemberState)

csThreshold ∷ !(Maybe UnitInterval)

csEpochNo ∷ !EpochNo

Current epoch number. This is necessary to interpret committee member states

data HotCredAuthStatus

Constructors

MemberAuthorized (Credential 'HotCommitteeRole)

MemberNotAuthorized
Member enacted, but no hot credential for voting has been registered

MemberResigned (Maybe Anchor)

data MemberStatus

Constructors

Active

Expired

Unrecognized
This can happen when a hot credential for an unknown cold credential exists. Such Committee member will be either removed from the state at the next epoch boundary or enacted as a new member.

data NextEpochChange

Constructors

ToBeEnacted

ToBeRemoved
Member will be removed

NoChangeExpected

ToBeExpired

TermAdjusted EpochNo

GetCurrentPParams

queryCurrentPParams ∷ EraGov era ⇒ NewEpochState era → PParams era

This is a simple lookup into the state for the values of current protocol parameters. These values can change on the epoch boundary. Use queryFuturePParams to see if we are aware of any upcoming changes.

GetFuturePParams

queryFuturePParams ∷ EraGov era ⇒ NewEpochState era → Maybe (PParams era)

This query will return values for protocol parameters that are likely to be adopted at the next epoch boundary. It is only when we passed 2 stability windows before the end of the epoch that users can rely on this query to produce stable results.

GetProposals

queryProposals

Arguments

∷ ConwayEraGov era

⇒ NewEpochState era

→ Set GovActionId
Specify a set of Governance Action IDs to filter the proposals. When this set is empty, all the proposals considered for ratification will be returned.

→ Seq (GovActionState era)

Query proposals that are considered for ratification.

GetRatifyState

queryRatifyState ∷ ConwayEraGov era ⇒ NewEpochState era → RatifyState era

Query ratification state.

GetStakePoolDefaultVote

queryStakePoolDefaultVote

Arguments

∷ (EraCertState era, ConwayEraAccounts era)

⇒ NewEpochState era

→ KeyHash 'StakePool
Specify the key hash of the pool whose default vote should be returned.

→ DefaultVote

Query a stake pool's reward account delegatee which determines the pool's default vote in absence of an explicit vote. Note that this is different from the delegatee determined by the credential of the stake pool itself.

data DefaultVote

Default vote that will be used for Stake Pool.

Constructors

DefaultNo
Reward account is delegated to a DRepKeyHash, DRepScriptHash or undelegated: default vote is No.

DefaultAbstain
Reward account is delegated to DRepAlwaysAbstain: default vote is Abstain, except for HardForkInitiation actions.

DefaultNoConfidence
Reward account is delegated to DRepAlwaysNoConfidence: default vote is Yes in case of a NoConfidence action, otherwise No.

GetPoolState

queryPoolParameters ∷ EraCertState era ⇒ NewEpochState era → Set (KeyHash 'StakePool) → Map (KeyHash 'StakePool) PoolParams

Query the current PoolParams.

queryPoolState ∷ EraCertState era ⇒ NewEpochState era → Maybe (Set (KeyHash 'StakePool)) → QueryPoolStateResult

Query the QueryPoolStateResult. This is slightly different from the internal representation used by Ledger and is intended to resemble how the internal representation used to be.

data QueryPoolStateResult

Used only for the queryPoolState query. This resembles the older way of representing StakePoolState in Ledger.

Constructors

QueryPoolStateResult

Fields

qpsrStakePoolParams ∷ !(Map (KeyHash 'StakePool) PoolParams)

qpsrFutureStakePoolParams ∷ !(Map (KeyHash 'StakePool) PoolParams)

qpsrRetiring ∷ !(Map (KeyHash 'StakePool) EpochNo)

qpsrDeposits ∷ !(Map (KeyHash 'StakePool) Coin)

mkQueryPoolStateResult ∷ (∀ x. Map (KeyHash 'StakePool) x → Map (KeyHash 'StakePool) x) → PState era → QueryPoolStateResult

For testing

getNextEpochCommitteeMembers ∷ ConwayEraGov era ⇒ NewEpochState era → Map (Credential 'ColdCommitteeRole) EpochNo