Cardano.Ledger.Api.State.Query.CommitteeMembersState

Documentation

data CommitteeMemberState c

Constructors

CommitteeMemberState

Fields

cmsHotCredAuthStatus ∷ !(HotCredAuthStatus c)

cmsStatus ∷ !MemberStatus

cmsExpiration ∷ !(Maybe EpochNo)

Absolute epoch number when the member expires

cmsNextEpochChange ∷ !NextEpochChange

Changes to the member at the next epoch

data CommitteeMembersState c

Constructors

CommitteeMembersState

Fields

csCommittee ∷ !(Map (Credential 'ColdCommitteeRole c) (CommitteeMemberState c))

csThreshold ∷ !(Maybe UnitInterval)

csEpochNo ∷ !EpochNo

Current epoch number. This is necessary to interpret committee member states

data HotCredAuthStatus c

Constructors

MemberAuthorized (Credential 'HotCommitteeRole c)

MemberNotAuthorized
Member enacted, but no hot credential for voting has been registered

MemberResigned (Maybe (Anchor c))

data MemberStatus

Constructors

Active

Expired

Unrecognized
This can happen when a hot credential for an unknown cold credential exists. Such Committee member will be either removed from the state at the next epoch boundary or enacted as a new member.

data NextEpochChange

Constructors

ToBeEnacted

ToBeRemoved
Member will be removed

NoChangeExpected

ToBeExpired

TermAdjusted EpochNo