Cardano.Ledger.Api.Transition

Description

This module is used for defining initial configuration for all. It is also used in testing and benchmarking to initilize a chain in a particular era without going through the trouble of generating all the history for preceeding eras.

Documentation

class (EraTxOut era, EraGov era, EraStake era, EraGenesis era, EraCertState era, Eq (TransitionConfig era), Show (TransitionConfig era), FromJSON (TransitionConfig era), Default (StashedAVVMAddresses era)) ⇒ EraTransition era

Register the initial information in the NewEpochState.

HERE BE DRAGONS! This interfaced is intended to help in testing.

In production, the genesis should not contain any initial information about accounts, stake pools or dreps.

Warning - Should only be used in testing and benchmarking. Will result in an error when NetworkId is set to Mainnet

Minimal complete definition

mkTransitionConfig, injectIntoTestState, tcPreviousEraConfigL, tcTranslationContextL

data family TransitionConfig era

Cumulative configuration that is needed to be able to start in a current era

mkLatestTransitionConfig ∷ ShelleyGenesis → AlonzoGenesis → ConwayGenesis → DijkstraGenesis → TransitionConfig LatestKnownEra

mkTransitionConfig

Arguments

∷ EraTransition era

⇒ TranslationContext era
Translation context necessary for advancing from previous era into the current one. This will usually be the contents of genesis file, if one exists for the current era

→ TransitionConfig (PreviousEra era)
Transition configuration for the previous era.

→ TransitionConfig era

mkShelleyTransitionConfig ∷ ShelleyGenesis → TransitionConfig ShelleyEra

Constructor for the base Shelley TransitionConfig

tcShelleyGenesisL ∷ EraTransition era ⇒ Lens' (TransitionConfig era) ShelleyGenesis

Lens for the ShelleyGenesis from the TransitionConfig. Default implementation looks in the previous era's config

tcPreviousEraConfigL ∷ EraTransition era ⇒ Lens' (TransitionConfig era) (TransitionConfig (PreviousEra era))

In case when a previous era is available, we should always be able to access TransitionConfig for the previous era, from within the current era's TransitionConfig

tcTranslationContextL ∷ EraTransition era ⇒ Lens' (TransitionConfig era) (TranslationContext era)

Lens for the TranslationContext for the current era from the TransitionConfig Translation context is a different name for the Genesis type for each era, they are one and the same concept.

Genesis

class Era era ⇒ EraGenesis era

Associated Types

type Genesis era

type Genesis era = NoGenesis era

data NoGenesis era

Constructors

NoGenesis

data ShelleyGenesis

Shelley genesis information

Note that this is needed only for a pure Shelley network, hence it being defined here rather than in its own module. In mainnet, Shelley will transition naturally from Byron, and thus will never have its own genesis information.

Constructors

ShelleyGenesis

Fields

sgSystemStart ∷ !UTCTime

sgNetworkMagic ∷ !Word32

sgNetworkId ∷ !Network

sgActiveSlotsCoeff ∷ !PositiveUnitInterval

sgSecurityParam ∷ !(NonZero Word64)

sgEpochLength ∷ !EpochSize

sgSlotsPerKESPeriod ∷ !Word64

sgMaxKESEvolutions ∷ !Word64

sgSlotLength ∷ !NominalDiffTimeMicro

sgUpdateQuorum ∷ !Word64

sgMaxLovelaceSupply ∷ !Word64

sgProtocolParams ∷ !(PParams ShelleyEra)

sgGenDelegs ∷ !(Map (KeyHash 'Genesis) GenDelegPair)

sgInitialFunds ∷ ListMap Addr Coin

sgInitialFunds is intentionally kept lazy, as it can otherwise cause out-of-memory problems in testing and benchmarking.

sgStaking ∷ ShelleyGenesisStaking

sgStaking is intentionally kept lazy, as it can otherwise cause out-of-memory problems in testing and benchmarking.

newtype AlonzoGenesis

All configuration that is necessary to bootstrap AlonzoEra from ShelleyGenesis

Constructors

AlonzoGenesisWrapper

Fields

unAlonzoGenesisWrapper ∷ UpgradeAlonzoPParams Identity

Bundled Patterns

pattern AlonzoGenesis ∷ CoinPerWord → CostModels → Prices → ExUnits → ExUnits → Natural → Natural → Natural → AlonzoGenesis

data ConwayGenesis

Constructors

ConwayGenesis

Fields

cgUpgradePParams ∷ !(UpgradeConwayPParams Identity)

cgConstitution ∷ !(Constitution ConwayEra)

cgCommittee ∷ !(Committee ConwayEra)

cgDelegs ∷ ListMap (Credential 'Staking) Delegatee

cgInitialDReps ∷ ListMap (Credential 'DRepRole) DRepState

newtype DijkstraGenesis

Constructors

DijkstraGenesis

Fields

dgUpgradePParams ∷ UpgradeDijkstraPParams Identity DijkstraEra

Functions for Testing

tcInitialPParamsG ∷ EraTransition era ⇒ SimpleGetter (TransitionConfig era) (PParams era)

Get the initial PParams for the current era from the TransitionConfig. Note that this is only useful for testing and fast forward functionality, because this function assumes no on-chain changes to PParams through PParamsUpdate functionality.

Default implementation will use the PParams from the Previous era and the current TranslationContext to construct PParams for the current era.

Warning - Should only be used in testing and benchmarking

tcInitialFundsL ∷ (HasCallStack, EraTransition era) ⇒ Lens' (TransitionConfig era) (ListMap Addr Coin)

Get the initial funds from the TransitionConfig. This value must be non-empty only during testing and benchmarking, it must never contain anything on a real system.

Warning - Should only be used in testing and benchmarking. Will result in an error when NetworkId is set to Mainnet

tcInitialStakingL ∷ (HasCallStack, EraTransition era) ⇒ Lens' (TransitionConfig era) ShelleyGenesisStaking

Get the initial staking from the TransitionConfig. This value must be non-empty only during testing and benchmarking, it must never contain anything on a real system.

Warning - Should only be used in testing and benchmarking. Will result in an error when NetworkId is set to Mainnet

createInitialState ∷ (EraTransition era, HasCallStack) ⇒ TransitionConfig era → NewEpochState era

Helper function for constructing the initial state for any era

Warning - Should only be used in testing and benchmarking. Will result in an error when NetworkId is set to Mainnet

This function does not register any initial funds or delegates.

injectIntoTestState

Arguments

∷ EraTransition era

⇒ TransitionConfig era
Extract data from the given transition configuration and store it in the given state.

Warning - Should only be used in testing and benchmarking. Will result in an error when NetworkId is set to Mainnet.

→ NewEpochState era

→ NewEpochState era