Cardano.Ledger.Api.Tx.Address

Address

data Addr c

An address for UTxO.

Contents of Addr data type are intentionally left as lazy, otherwise operating on compact form of an address will result in redundant work.

Constructors

Addr Network (PaymentCredential c) (StakeReference c)

AddrBootstrap (BootstrapAddress c)

getNetwork ∷ Addr c → Network

Lookup a Network Id for an Address

newtype BootstrapAddress c

Constructors

BootstrapAddress

Fields

unBootstrapAddress ∷ Address

serialiseAddr ∷ Addr c → ByteString

Serialise an address to the external format.

Strict decoders

Decoders below will only decode addresses that are allowed to be placed on chain today. Historically there were a few bugs in the decoder which allowed a few malformed addressed to be placed on chain. If you need backwards compatibility, reach out for decodeAddrLenient.

decodeAddr ∷ (Crypto c, MonadFail m) ⇒ ByteString → m (Addr c)

Strict decoder for an address from a ByteString. This will not let you decode some of the buggy addresses that have been placed on chain. This decoder is intended for addresses that are to be placed on chian today.

decodeAddrEither ∷ Crypto c ⇒ ByteString → Either String (Addr c)

Same as decodeAddr, but produces an Either result

decodeAddrShort ∷ (Crypto c, MonadFail m) ⇒ ShortByteString → m (Addr c)

Same as decodeAddr, but works on ShortByteString

decodeAddrShortEither ∷ Crypto c ⇒ ShortByteString → Either String (Addr c)

Same as decodeAddrShort, but produces an Either result

Lenient decoders

These lenient decoders do not fail for addresses with known bugs

data DecAddr c

Decoded Address.

Constructors

DecAddr (Addr c)
Address was decoded with no problems

DecAddrBadPtr (Addr c)
Address was decoded, but it contains an invalid Ptr

DecAddrUnconsumed
Address was decoded, but not all of input was consumed

Fields

(Addr c)

ByteString

Left over bytes after consuming the input

decodeAddrLenient ∷ (Crypto c, MonadFail m) ⇒ ByteString → m (Addr c)

This is a lenient decoder that will disregard known bugs in the address deserialization. This function is intended for clients that need to deal with historical data that has already been placed on chain. If you also require information on what exactly is bad in the address, or you would like to guard only against a specific bug, you should use decodeAddrLenientEither instead.

Since: 1.8.0

decodeAddrLenientEither ∷ Crypto c ⇒ ByteString → Either String (DecAddr c)

Decode an address and fail only for addresses that could have never been placed on chain, while decoding addresses with information about potential problems in them. Similar to decodeAddrLenient, this function is not intended for addresses that will be placed into a new transaction.

Since: 1.8.0

Reward Account

data RewardAccount c

An account based address for rewards

Constructors

RewardAccount

Fields

raNetwork ∷ !Network

raCredential ∷ !(Credential 'Staking c)

Bundled Patterns

pattern RewardAcnt ∷ Network → Credential 'Staking c → RewardAccount c

mkRwdAcnt ∷ Network → Credential 'Staking c → RewardAcnt c

serialiseRewardAccount ∷ RewardAcnt c → ByteString

Serialise a reward account to the external format.

deserialiseRewardAccount ∷ Crypto c ⇒ ByteString → Maybe (RewardAcnt c)

Deserialise a reward account from the external format. This will fail if the input data is not in the right format (or if there is trailing data).

Deprecations

type RewardAcnt = RewardAccount

serialiseRewardAcnt ∷ RewardAcnt c → ByteString

deserialiseRewardAcnt ∷ Crypto c ⇒ ByteString → Maybe (RewardAcnt c)