Cardano.Ledger.Api.Tx.AuxData

Documentation

class (Era era, Eq (TxAuxData era), EqRaw (TxAuxData era), Show (TxAuxData era), NoThunks (TxAuxData era), ToCBOR (TxAuxData era), EncCBOR (TxAuxData era), DecCBOR (Annotator (TxAuxData era)), HashAnnotated (TxAuxData era) EraIndependentTxAuxData) ⇒ EraTxAuxData era

TxAuxData which may be attached to a transaction

Minimal complete definition

mkBasicTxAuxData, metadataTxAuxDataL, validateTxAuxData

Associated Types

type TxAuxData era = (r ∷ Type) | r → era

mkBasicTxAuxData ∷ EraTxAuxData era ⇒ TxAuxData era

metadataTxAuxDataL ∷ EraTxAuxData era ⇒ Lens' (TxAuxData era) (Map Word64 Metadatum)

hashTxAuxData ∷ EraTxAuxData era ⇒ TxAuxData era → TxAuxDataHash

Compute a hash of TxAuxData

validateTxAuxData ∷ EraTxAuxData era ⇒ ProtVer → TxAuxData era → Bool

Shelley

newtype ShelleyTxAuxData era

Constructors

MkShelleyTxAuxData (MemoBytes (ShelleyTxAuxDataRaw era))

Bundled Patterns

pattern ShelleyTxAuxData ∷ Era era ⇒ Map Word64 Metadatum → ShelleyTxAuxData era

data Metadatum

A generic metadatum type.

Constructors

Map ![(Metadatum, Metadatum)]

List ![Metadatum]

I !Integer

B !ByteString

S !Text

Allegra

class EraTxAuxData era ⇒ AllegraEraTxAuxData era

Minimal complete definition

nativeScriptsTxAuxDataL

nativeScriptsTxAuxDataL ∷ AllegraEraTxAuxData era ⇒ Lens' (TxAuxData era) (StrictSeq (NativeScript era))

newtype AllegraTxAuxData era

Constructors

MkAlegraTxAuxData (MemoBytes (AllegraTxAuxDataRaw era))

Bundled Patterns

pattern AllegraTxAuxData ∷ (Era era, EncCBOR (NativeScript era)) ⇒ Map Word64 Metadatum → StrictSeq (NativeScript era) → AllegraTxAuxData era

Alonzo

class AllegraEraTxAuxData era ⇒ AlonzoEraTxAuxData era

Minimal complete definition

plutusScriptsTxAuxDataL

plutusScriptsTxAuxDataL ∷ AlonzoEraTxAuxData era ⇒ Lens' (TxAuxData era) (Map Language (NonEmpty PlutusBinary))

newtype AlonzoTxAuxData era

Constructors

MkAlonzoTxAuxData (MemoBytes (AlonzoTxAuxDataRaw era))

Bundled Patterns

pattern AlonzoTxAuxData ∷ (HasCallStack, AlonzoEraScript era) ⇒ Map Word64 Metadatum → StrictSeq (NativeScript era) → Map Language (NonEmpty PlutusBinary) → AlonzoTxAuxData era
Construct auxiliary data. Make sure not to supply plutus script versions that are not supported in this era, because it will result in a runtime exception. Use mkAlonzoTxAuxData instead if you need runtime safety guarantees.

pattern AlonzoTxAuxData' ∷ Map Word64 Metadatum → StrictSeq (NativeScript era) → Map Language (NonEmpty PlutusBinary) → AlonzoTxAuxData era

mkAlonzoTxAuxData ∷ (Foldable f, AlonzoEraScript era) ⇒ Map Word64 Metadatum → f (AlonzoScript era) → AlonzoTxAuxData era

Helper function that will construct Auxiliary data from Metadatum map and a list of scripts.

Note that the relative order of same type scripts will be preserved.

getAlonzoTxAuxDataScripts ∷ AlonzoEraScript era ⇒ AlonzoTxAuxData era → StrictSeq (AlonzoScript era)

Upgrade

binaryUpgradeTxAuxData ∷ (Era era, ToCBOR (TxAuxData (PreviousEra era)), DecCBOR (Annotator (TxAuxData era))) ⇒ TxAuxData (PreviousEra era) → Except DecoderError (TxAuxData era)

Translate tx auxData through its binary representation from previous to current era.

upgradeTxAuxData ∷ (EraApi era, EraTxAuxData (PreviousEra era)) ⇒ TxAuxData (PreviousEra era) → TxAuxData era

Upgrade txAuxData from the previous era. Warning - This may not preserve the underlying binary representation. Use binaryUpgradeTxAuxData instead, if you need to preserve the serialised form.