Cardano.Ledger.Api.Tx.In

Transaction input

data TxIn c

The input of a UTxO.

Constructors

TxIn !(TxId c) !TxIx

Transaction ID

newtype TxId c

A unique ID of a transaction, which is computable from the transaction.

Constructors

TxId

Fields

unTxId ∷ SafeHash c EraIndependentTxBody

mkTxInPartial ∷ HasCallStack ⇒ TxId c → Integer → TxIn c

Construct TxIn while throwing an error for an out of range TxIx. Make sure to use it only for testing.

Transaction index

data TxIx

Transaction index.

mkTxIx ∷ Word16 → TxIx

Construct a TxIx from a 16 bit unsigned integer

txIxToInt ∷ TxIx → Int

txIxFromIntegral ∷ (Integral a, MonadFail m) ⇒ a → m TxIx

mkTxIxPartial ∷ HasCallStack ⇒ Integer → TxIx

Construct a TxIx from an arbitrary precision Integer. Throws an error for values out of range. Make sure to use it only for testing.