Cardano.Ledger.Api.Tx.Wits

Shelley onwards

class (EraScript era, Eq (TxWits era), EqRaw (TxWits era), Show (TxWits era), Monoid (TxWits era), NoThunks (TxWits era), ToCBOR (TxWits era), EncCBOR (TxWits era), DecCBOR (Annotator (TxWits era))) ⇒ EraTxWits era

A collection of witnesses in a Tx

Minimal complete definition

addrTxWitsL, bootAddrTxWitsL, scriptTxWitsL, upgradeTxWits

Associated Types

type TxWits era = (r ∷ Type) | r → era

mkBasicTxWits ∷ EraTxWits era ⇒ TxWits era

Address witness

addrTxWitsL ∷ EraTxWits era ⇒ Lens' (TxWits era) (Set (WitVKey 'Witness))

data KeyRole

The role of a key.

All key roles are fixed and unique, except for the Witness role. In particular, keys can be cast to a Witness role with the help of asWitness, because same witness can be valid for many roles.

In fact, it is perfectly allowable for a key to be used in many roles by the end user; there is nothing prohibiting somebody using the same underlying key or a script as their payment and staking credential, as well as the key for their stake pool. However, in the ledger code mixing up keys with different roles could be catastrophic, that is why we have this separation.

Constructors

Witness

WitVKey

data WitVKey (kr ∷ KeyRole) where

Proof/Witness that a transaction is authorized by the given key holder.

Bundled Patterns

pattern WitVKey ∷ Typeable kr ⇒ VKey kr → SignedDSIGN DSIGN (Hash HASH EraIndependentTxBody) → WitVKey kr

witVKeyBytes ∷ ∀ (kr ∷ KeyRole). WitVKey kr → ByteString

Access CBOR encoded representation of the witness. Evaluated lazily

witVKeyHash ∷ ∀ (kr ∷ KeyRole). WitVKey kr → KeyHash 'Witness

Access computed hash. Evaluated lazily

Byron address witness

bootAddrTxWitsL ∷ EraTxWits era ⇒ Lens' (TxWits era) (Set BootstrapWitness)

data BootstrapWitness

Script witness

scriptTxWitsL ∷ EraTxWits era ⇒ Lens' (TxWits era) (Map ScriptHash (Script era))

hashScriptTxWitsL ∷ EraTxWits era ⇒ Lens (TxWits era) (TxWits era) (Map ScriptHash (Script era)) [Script era]

This is a helper lens that will hash the scripts when adding as witnesses.

Alonzo onwards

class (EraTxWits era, AlonzoEraScript era) ⇒ AlonzoEraTxWits era

Minimal complete definition

datsTxWitsL, rdmrsTxWitsL

datsTxWitsL ∷ AlonzoEraTxWits era ⇒ Lens' (TxWits era) (TxDats era)

hashDataTxWitsL ∷ AlonzoEraTxWits era ⇒ Lens (TxWits era) (TxWits era) (TxDats era) [Data era]

This is a convenience Lens that will hash the Data when it is being added to the TxWits. See datsTxWitsL for a version that aloows setting TxDats instead.

data TxDats era where

Note that TxDats are based on MemoBytes since we must preserve the original bytes for the ScriptIntegrity. Since the TxDats exist outside of the transaction body, this is how we ensure that they are not manipulated.

Bundled Patterns

pattern TxDats ∷ Era era ⇒ Map DataHash (Data era) → TxDats era

pattern TxDats' ∷ Map DataHash (Data era) → TxDats era

unTxDats ∷ TxDats era → Map DataHash (Data era)

rdmrsTxWitsL ∷ AlonzoEraTxWits era ⇒ Lens' (TxWits era) (Redeemers era)

data Redeemers era where

Note that Redeemers are based on MemoBytes since we must preserve the original bytes for the ScriptIntegrity. Since the Redeemers exist outside of the transaction body, this is how we ensure that they are not manipulated.

Bundled Patterns

pattern Redeemers ∷ AlonzoEraScript era ⇒ Map (PlutusPurpose AsIx era) (Data era, ExUnits) → Redeemers era

unRedeemers ∷ Redeemers era → Map (PlutusPurpose AsIx era) (Data era, ExUnits)

type family PlutusPurpose (f ∷ Type → Type → Type) era = (r ∷ Type) | r → era

data AlonzoPlutusPurpose (f ∷ Type → Type → Type) era

Constructors

AlonzoSpending !(f Word32 TxIn)

AlonzoMinting !(f Word32 PolicyID)

AlonzoCertifying !(f Word32 (TxCert era))

AlonzoRewarding !(f Word32 RewardAccount)

newtype AsIx ix it

Constructors

AsIx

Fields

unAsIx ∷ ix

newtype AsItem ix it

Constructors

AsItem

Fields

unAsItem ∷ it

Conway

data ConwayPlutusPurpose (f ∷ Type → Type → Type) era

Constructors

ConwaySpending !(f Word32 TxIn)

ConwayMinting !(f Word32 PolicyID)

ConwayCertifying !(f Word32 (TxCert era))

ConwayRewarding !(f Word32 RewardAccount)

ConwayVoting !(f Word32 Voter)

ConwayProposing !(f Word32 (ProposalProcedure era))