Test.Cardano.Ledger.Api.Examples.Consensus.Shelley

Documentation

data ShelleyResultExamples era

Constructors

ShelleyResultExamples

Fields

srePParams ∷ PParams era

sreProposedPPUpdates ∷ ProposedPPUpdates era

srePoolDistr ∷ PoolDistr

sreNonMyopicRewards ∷ Map (Either Coin (Credential 'Staking)) (Map (KeyHash 'StakePool) Coin)

sreShelleyGenesis ∷ ShelleyGenesis

data ShelleyLedgerExamples era

Constructors

ShelleyLedgerExamples

Fields

sleBlock ∷ Block (BHeader StandardCrypto) era

sleHashHeader ∷ HashHeader

sleTx ∷ Tx era

sleApplyTxError ∷ ApplyTxError era

sleRewardsCredentials ∷ Set (Either Coin (Credential 'Staking))

sleResultExamples ∷ ShelleyResultExamples era

sleNewEpochState ∷ NewEpochState era

sleChainDepState ∷ ChainDepState

sleTranslationContext ∷ TranslationContext era

defaultShelleyLedgerExamples ∷ ∀ era. (EraSegWits era, EraGov era, EraStake era, EraCertState era, PredicateFailure (EraRule "DELEGS" era) ~ ShelleyDelegsPredFailure era, PredicateFailure (EraRule "LEDGER" era) ~ ShelleyLedgerPredFailure era, Default (StashedAVVMAddresses era), ProtVerAtMost era 4) ⇒ (TxBody era → [KeyPair 'Witness] → TxWits era) → Value era → TxBody era → TxAuxData era → TranslationContext era → ShelleyLedgerExamples era

exampleShelleyLedgerBlock ∷ ∀ era. EraSegWits era ⇒ Tx era → Block (BHeader StandardCrypto) era

exampleHashHeader ∷ HashHeader

exampleTx ∷ ∀ era. EraTx era ⇒ (TxBody era → [KeyPair 'Witness] → TxWits era) → TxBody era → TxAuxData era → Tx era

This is not a valid transaction. We don't care, we are only interested in serialisation, not validation.

exampleProposedPParamsUpdates ∷ EraPParams era ⇒ ProposedPPUpdates era

examplePoolDistr ∷ PoolDistr

exampleNonMyopicRewards ∷ Map (Either Coin (Credential 'Staking)) (Map (KeyHash 'StakePool) Coin)

exampleNewEpochState ∷ ∀ era. (EraTxOut era, EraGov era, EraStake era, EraCertState era, Default (StashedAVVMAddresses era)) ⇒ Value era → PParams era → PParams era → NewEpochState era

This is probably not a valid ledger. We don't care, we are only interested in serialisation, not validation.

exampleLedgerChainDepState ∷ Word64 → ChainDepState

exampleCoin ∷ Coin

exampleTxBodyShelley ∷ TxBody ShelleyEra

exampleAuxDataMap ∷ Map Word64 Metadatum

exampleTxIns ∷ Set TxIn

exampleCerts ∷ (ShelleyEraTxCert era, ProtVerAtMost era 8) ⇒ StrictSeq (TxCert era)

exampleWithdrawals ∷ Withdrawals

exampleProposedPPUpdates ∷ EraPParams era ⇒ ProposedPPUpdates era

examplePayKey ∷ KeyPair 'Payment

exampleStakeKey ∷ KeyPair 'Staking

exampleKeys ∷ AllIssuerKeys StandardCrypto r

exampleAuxiliaryDataShelley ∷ TxAuxData ShelleyEra

examplePoolParams ∷ PoolParams

ledgerExamplesShelley ∷ ShelleyLedgerExamples ShelleyEra

ShelleyLedgerExamples for Shelley era

testShelleyGenesis ∷ ShelleyGenesis

These are dummy values.

helper functions used in other examples

keyToCredential ∷ KeyPair r → Credential r

mkKeyHash ∷ ∀ discriminator. Int → KeyHash discriminator

mkWitnessesPreAlonzo ∷ EraTx era ⇒ Proxy era → TxBody era → [KeyPair 'Witness] → ShelleyTxWits era

mkScriptHash ∷ Int → ScriptHash

from Translation

emptyFromByronTranslationContext ∷ FromByronTranslationContext

Trivial FromByronTranslationContext value, for use in cases where we do not need to translate from Byron to Shelley.