Cardano.Ledger.Api.Era

Eras

class (Typeable era, KnownNat (ProtVerLow era), KnownNat (ProtVerHigh era), ProtVerLow era <= ProtVerHigh era, MinVersion <= ProtVerLow era, MinVersion <= ProtVerHigh era, CmpNat (ProtVerLow era) MaxVersion ~ 'LT, CmpNat (ProtVerHigh era) MaxVersion ~ 'LT, ProtVerLow era <= MaxVersion, ProtVerHigh era <= MaxVersion) ⇒ Era era

Minimal complete definition

eraName

Associated Types

type PreviousEra era = (r ∷ Type) | r → era

Map an era to its predecessor.

For example:

type instance PreviousEra (AllegraEra c) = ShelleyEra c

type ProtVerLow era ∷ Nat

Lowest major protocol version for this era

type ProtVerHigh era ∷ Nat

Highest major protocol version for this era. By default se to ProtVerLow

type ProtVerHigh era = ProtVerLow era

eraName ∷ Era era ⇒ String

Textual name of the current era.

Designed to be used with TypeApplications:

>>> eraName @ByronEra
Byron

Byron

data ByronEra

This is the era that preceded Shelley era. It cannot have any other class instances, except for Era type class.

Shelley

type Shelley = ShelleyEra

data ShelleyEra

Allegra

type Allegra = AllegraEra

data AllegraEra

The Allegra era

Mary

type Mary = MaryEra

data MaryEra

Alonzo

type Alonzo = AlonzoEra

data AlonzoEra

The Alonzo era

Babbage

type Babbage = BabbageEra

data BabbageEra

The Babbage era

Conway

type Conway = ConwayEra

data ConwayEra

The Conway era

Latest Known

type LatestKnownEra = ConwayEra

Sometimes it is useful to specify that a type corresponds to a latest era that is currently implemented

Protocol version

Value level

eraProtVerHigh ∷ Era era ⇒ Version

Get the value level Version of the highest major protocol version for the supplied era.

eraProtVerLow ∷ Era era ⇒ Version

Get the value level Version of the lowest major protocol version for the supplied era.

Type level constraints

type AtLeastEra atLeastEra era = ProtVerAtLeast era (ProtVerLow atLeastEra)

Restrict the era to equal to atLeastEra or come after it

type AtMostEra eraMostEra era = ProtVerAtMost era (ProtVerHigh eraMostEra)

Restrict the era to equal to eraName or come before it.

type ExactEra inEra era = ProtVerInBounds era (ProtVerLow inEra) (ProtVerHigh inEra)

Restrict an era to the specific era through the protocol version. This is equivalent to (inEra (Crypto era) ~ era)

type family ProtVerAtLeast era (l ∷ Nat) where ...

Requirement for the era's highest protocol version to be higher or equal to the supplied value

Equations

ProtVerAtLeast era l = ProtVerIsInBounds "at least" era l (l <=? ProtVerHigh era)

type family ProtVerAtMost era (h ∷ Nat) where ...

Requirement for the era's lowest protocol version to be lower or equal to the supplied value

Equations

ProtVerAtMost era h = ProtVerIsInBounds "at most" era h (ProtVerLow era <=? h)

type ProtVerInBounds era (l ∷ Nat) (h ∷ Nat) = (ProtVerAtLeast era l, ProtVerAtMost era h)

Restrict a lower and upper bounds of the protocol version for the particular era

atLeastEra ∷ AtLeastEra eraName era ⇒ ()

Enforce era to be at least the specified era at the type level. In other words compiler will produce type error when applied to eras prior to the specified era. This function should be used in order to avoid redundant constraints warning.

For example these will type check

>>> atLeastEra @BabbageEra @ConwayEra
>>> atLeastEra @BabbageEra @BabbageEra

However this will result in a type error

>>> atLeastEra @BabbageEra @AlonzoEra

atMostEra ∷ AtMostEra eraName era ⇒ ()

Enforce era to be at most the specified era at the type level. In other words compiler will produce type error when applied to eras prior to the specified era. This function should be used in order to avoid redundant constraints warning.

For example these will type check

>>> atMostEra @BabbageEra @ShelleyEra
>>> atMostEra @AlonzoEra @MaryEra

However this will result in a type error

>>> atMostEra @BabbageEra @ConwayEra