Cardano.Ledger.Api.Era

Eras

class (Typeable era, KnownNat (ProtVerLow era), KnownNat (ProtVerHigh era), ProtVerLow era <= ProtVerHigh era, MinVersion <= ProtVerLow era, MinVersion <= ProtVerHigh era, CmpNat (ProtVerLow era) MaxVersion ~ 'LT, CmpNat (ProtVerHigh era) MaxVersion ~ 'LT) ⇒ Era era

Associated Types

type EraName era ∷ Symbol

type PreviousEra era = (r ∷ Type) | r → era

Map an era to its predecessor.

For example:

type instance PreviousEra AllegraEra = ShelleyEra

type ProtVerLow era ∷ Nat

Lowest major protocol version for this era

type ProtVerHigh era ∷ Nat

Highest major protocol version for this era. By default se to ProtVerLow

type ProtVerHigh era = ProtVerLow era

class (EraTx era, EraTxOut era, EraTxBody era, EraTxAuxData era, EraTxWits era, EraScript era, EraPParams era, EraBlockBody era, EraTxCert era) ⇒ EraApi era where

Associated Types

type TxUpgradeError era ∷ Type

Upgrade transactions from the previous era. This will preseve the values in corresponding Haskell structures where possible, however it will not preserve the binary representation for memoized types. If binary representation and validity of signatures, scriptIntegrityHash and auxDataHash must be retained, then the corresponding binaryUpgradeTx, binaryUpgradeTxBody, binaryUpgradeTxAuxData and binaryUpgradeTxWits functions should be used instead. Compare the two types of upgrade: - `upgrade[Tx, TxBody, TxAuxData, TxWits]` will use the Haskell representation, but will not preserve the serialised form. However, they will be suitable for iterated translation through eras. - `binaryUpgrade[Tx, TxBody, TxAuxData, TxWits]` will preserve the binary representation, but are not guaranteed to work through multiple eras - that is, the serialised representation from era n is guaranteed valid in era n + 1, but not necessarily in era n + 2.

See below an example of how the upgrade function can change the underlying serialization:

>>> import Cardano.Ledger.Api.Era (BabbageEra, ConwayEra)
>>> import Test.QuickCheck
>>> import Test.Cardano.Ledger.Babbage.Arbitrary ()
>>> import Cardano.Ledger.Plutus.Data
>>> import Cardano.Ledger.Alonzo.TxWits (unTxDats)
>>> import Cardano.Ledger.MemoBytes (getMemoRawBytes)
>>> import Cardano.Ledger.Binary (serialize)

Let's generate an arbitrary Babbage TxWits: >>> witsBabbage <- generate $ arbitrary @(TxWits BabbageEra)

Now we upgrade it to Conway using the upgrade functionality in this module: >>> witsUpgraded = upgradeTxWits witsBabbage

We can check that the values in the data structures have been preserved. For simplicity, let's check the txDats: let dats = Map.map unData . unTxDats . txdats in dats witsBabbage == dats witsUpgraded True

However, the upgraded value will have a different serialized binary representation than the original, because the serialization of TxWits changed from Babbage to Conway >>> getMemoRawBytes witsBabbage == getMemoRawBytes witsUpgraded False >>> serialize (eraProtVerHigh BabbageEra) witsBabbage == serialize (eraProtVerHigh BabbageEra) witsUpgraded False

type TxUpgradeError era = Void

type TxBodyUpgradeError era ∷ Type

type TxBodyUpgradeError era = Void

Methods

upgradeTx ∷ EraTx (PreviousEra era) ⇒ Tx (PreviousEra era) → Either (TxUpgradeError era) (Tx era)

Upgrade a transaction from the previous era. Warning - This may not preserve the underlying binary representation. Use binaryUpgradeTx instead, if you need to preserve the serialised form.

upgradeTxBody ∷ EraTxBody (PreviousEra era) ⇒ TxBody (PreviousEra era) → Either (TxBodyUpgradeError era) (TxBody era)

Upgrade a transaction body from the previous era. Warning - This may not preserve the underlying binary representation. Use binaryUpgradeTxBody instead, if you need to preserve the serialised form.

upgradeTxAuxData ∷ EraTxAuxData (PreviousEra era) ⇒ TxAuxData (PreviousEra era) → TxAuxData era

Upgrade txAuxData from the previous era. Warning - This may not preserve the underlying binary representation. Use binaryUpgradeTxAuxData instead, if you need to preserve the serialised form.

upgradeTxWits ∷ EraTxWits (PreviousEra era) ⇒ TxWits (PreviousEra era) → TxWits era

Upgrade txWits from the previous era. Warning - This may not preserve the underlying binary representation. Use binaryUpgradeTxWits instead, if you need to preserve the serialised form.

upgradeNativeScript ∷ NativeScript (PreviousEra era) → NativeScript era

Upgrade a native script from the previous era.

eraName ∷ Era era ⇒ String

Textual name of the current era.

Designed to be used with TypeApplications:

>>> eraName @ByronEra
"Byron"

class (KnownSymbol eraName, Era (EraFromName eraName), EraName (EraFromName eraName) ~ eraName) ⇒ EraHasName (eraName ∷ Symbol)

This class exists in order to be able to derive the protocol version range for an era from its name. It is achieved by the means of injective definition of an era type from its name. This effectively closes the world for eras to only the ones that are deined in this module, however it does have to be that way. In other words, if there is ever a need to define custom eras that build on top of existing ledger eras, then we will need to remove injectivity from this type class as well as from the PreviousEra type family.

Associated Types

type EraFromName (eraName ∷ Symbol) = (era ∷ Type) | era → eraName

Byron

data ByronEra

This is the era that preceded Shelley era. It cannot have any other class instances, except for Era type class.

Shelley

data ShelleyEra

Allegra

data AllegraEra

Mary

data MaryEra

Alonzo

data AlonzoEra

Babbage

data BabbageEra

Conway

data ConwayEra

Dijkstra

data DijkstraEra

Latest Known

type LatestKnownEra = DijkstraEra

Sometimes it is useful to specify that a type corresponds to a latest era that is currently implemented

Protocol version

Value level

eraProtVerHigh ∷ Era era ⇒ Version

Get the value level Version of the highest major protocol version for the supplied era.

eraProtVerLow ∷ Era era ⇒ Version

Get the value level Version of the lowest major protocol version for the supplied era.

Type level constraints

type AtLeastEra (atLeastEra ∷ Symbol) era = ProtVerAtLeast era (ProtVerLow (EraFromName atLeastEra))

Restrict the era to equal to atLeastEra or come after it

type AtMostEra (eraMostEra ∷ Symbol) era = ProtVerAtMost era (ProtVerHigh (EraFromName eraMostEra))

Restrict the era to equal to eraName or come before it.

type ExactEra inEra era = ProtVerInBounds era (ProtVerLow inEra) (ProtVerHigh inEra)

Restrict an era to the specific era through the protocol version. This is equivalent to (inEra (Crypto era) ~ era)

type family ProtVerAtLeast era (l ∷ Nat) where ...

Requirement for the era's highest protocol version to be higher or equal to the supplied value

Equations

ProtVerAtLeast era l = ProtVerIsInBounds "at least" era l (l <=? ProtVerHigh era)

type family ProtVerAtMost era (h ∷ Nat) where ...

Requirement for the era's lowest protocol version to be lower or equal to the supplied value

Equations

ProtVerAtMost era h = ProtVerIsInBounds "at most" era h (ProtVerLow era <=? h)

type ProtVerInBounds era (l ∷ Nat) (h ∷ Nat) = (ProtVerAtLeast era l, ProtVerAtMost era h)

Restrict a lower and upper bounds of the protocol version for the particular era

atLeastEra ∷ ∀ (eraName ∷ Symbol) era. AtLeastEra eraName era ⇒ ()

Enforce era to be at least the specified era at the type level. In other words compiler will produce type error when applied to eras prior to the specified era. This function should be used in order to avoid redundant constraints warning.

For example these will type check

atLeastEra @"Babbage" @ConwayEra
atLeastEra @"Babbage" @BabbageEra

However this will result in a type error

atLeastEra @"Babbage" @AlonzoEra

atMostEra ∷ ∀ (eraName ∷ Symbol) era. AtMostEra eraName era ⇒ ()

Enforce era to be at most the specified era at the type level. In other words compiler will produce type error when applied to eras prior to the specified era. This function should be used in order to avoid redundant constraints warning.

For example these will type check

atMostEra @BabbageEra @ShelleyEra
atMostEra @AlonzoEra @MaryEra

However this will result in a type error

atMostEra @BabbageEra @ConwayEra