Cardano.Ledger.Api.Era

Eras

class (Crypto (EraCrypto era), Typeable era, KnownNat (ProtVerLow era), KnownNat (ProtVerHigh era), ProtVerLow era <= ProtVerHigh era, MinVersion <= ProtVerLow era, MinVersion <= ProtVerHigh era, CmpNat (ProtVerLow era) MaxVersion ~ 'LT, CmpNat (ProtVerHigh era) MaxVersion ~ 'LT, ProtVerLow era <= MaxVersion, ProtVerHigh era <= MaxVersion) => Era era

Minimal complete definition

eraName

Associated Types

type PreviousEra era = (r :: Type) | r -> era

type ProtVerLow era :: Nat

type ProtVerHigh era :: Nat

type ProtVerHigh era = ProtVerLow era

eraName :: Era era => String

Byron

data ByronEra c

Shelley

type Shelley = ShelleyEra StandardCrypto

data ShelleyEra c

Allegra

type Allegra = AllegraEra StandardCrypto

data AllegraEra c

Mary

type Mary = MaryEra StandardCrypto

data MaryEra era

Alonzo

type Alonzo = AlonzoEra StandardCrypto

data AlonzoEra c

Babbage

type Babbage = BabbageEra StandardCrypto

data BabbageEra c

Conway

type Conway = ConwayEra StandardCrypto

data ConwayEra c

Latest Known

type LatestKnownEra c = ConwayEra c

Sometimes it is useful to specify that a type corresponds to a latest era that is currently implemented

Crypto

data StandardCrypto

class (HashAlgorithm (HASH c), HashAlgorithm (ADDRHASH c), DSIGNAlgorithm (DSIGN c), KESAlgorithm (KES c), VRFAlgorithm (VRF c), ContextDSIGN (DSIGN c) ~ (), ContextKES (KES c) ~ (), ContextVRF (VRF c) ~ (), Typeable c) => Crypto c

Associated Types

type HASH c

type ADDRHASH c

type DSIGN c

type KES c

type VRF c

Protocol version

Value level

eraProtVerHigh :: Era era => Version

eraProtVerLow :: Era era => Version

Type level constraints

type AtLeastEra (eraName :: Type -> Type) era = ProtVerAtLeast era (ProtVerLow (eraName (EraCrypto era)))

type AtMostEra (eraName :: Type -> Type) era = ProtVerAtMost era (ProtVerHigh (eraName (EraCrypto era)))

type ExactEra (inEra :: Type -> Type) era = ProtVerInBounds era (ProtVerLow (inEra (EraCrypto era))) (ProtVerHigh (inEra (EraCrypto era)))

type family ProtVerAtLeast era (l :: Nat) where ...

Equations

ProtVerAtLeast era l = ProtVerIsInBounds "at least" era l (l <=? ProtVerHigh era)

type family ProtVerAtMost era (h :: Nat) where ...

Equations

ProtVerAtMost era h = ProtVerIsInBounds "at most" era h (ProtVerLow era <=? h)

type ProtVerInBounds era (l :: Nat) (h :: Nat) = (ProtVerAtLeast era l, ProtVerAtMost era h)

atLeastEra :: forall (eraName :: Type -> Type) era. AtLeastEra eraName era => ()

atMostEra :: forall (eraName :: Type -> Type) era. AtMostEra eraName era => ()