Cardano.Ledger.Api.Era

Eras

class (Crypto (EraCrypto era), Typeable era, KnownNat (ProtVerLow era), KnownNat (ProtVerHigh era), ProtVerLow era <= ProtVerHigh era, MinVersion <= ProtVerLow era, MinVersion <= ProtVerHigh era, CmpNat (ProtVerLow era) MaxVersion ~ 'LT, CmpNat (ProtVerHigh era) MaxVersion ~ 'LT, ProtVerLow era <= MaxVersion, ProtVerHigh era <= MaxVersion) ⇒ Era era

Associated Types

type EraCrypto era

type PreviousEra era = (r ∷ Type) | r → era

Map an era to its predecessor.

For example:

type instance PreviousEra (AllegraEra c) = ShelleyEra c

type ProtVerLow era ∷ Nat

Lowest major protocol version for this era

type ProtVerHigh era ∷ Nat

Highest major protocol version for this era. By default se to ProtVerLow

type ProtVerHigh era = ProtVerLow era

Byron

data ByronEra c

This is the era that preceded Shelley era. It cannot have any other class instances, except for Era type class.

Shelley

type Shelley = ShelleyEra StandardCrypto

data ShelleyEra c

Allegra

type Allegra = AllegraEra StandardCrypto

data AllegraEra c

The Allegra era

Mary

type Mary = MaryEra StandardCrypto

data MaryEra era

Alonzo

type Alonzo = AlonzoEra StandardCrypto

data AlonzoEra c

The Alonzo era

Babbage

type Babbage = BabbageEra StandardCrypto

data BabbageEra c

The Babbage era

Conway

type Conway = ConwayEra StandardCrypto

data ConwayEra c

The Conway era

Crypto

data StandardCrypto

The same crypto used on the net

class (HashAlgorithm (HASH c), HashAlgorithm (ADDRHASH c), DSIGNAlgorithm (DSIGN c), KESAlgorithm (KES c), VRFAlgorithm (VRF c), ContextDSIGN (DSIGN c) ~ (), ContextKES (KES c) ~ (), ContextVRF (VRF c) ~ (), Typeable c) ⇒ Crypto c

Associated Types

type HASH c

type ADDRHASH c

type DSIGN c

type KES c

type VRF c

Protocol version

Value level

eraProtVerHigh ∷ Era era ⇒ Version

Get the value level Version of the highest major protocol version for the supplied era.

eraProtVerLow ∷ Era era ⇒ Version

Get the value level Version of the lowest major protocol version for the supplied era.

Type level constraints

type AtLeastEra (eraName ∷ Type → Type) era = ProtVerAtLeast era (ProtVerLow (eraName (EraCrypto era)))

Restrict the era to equal to eraName or come after it

type AtMostEra (eraName ∷ Type → Type) era = ProtVerAtMost era (ProtVerHigh (eraName (EraCrypto era)))

Restrict the era to equal to eraName or come before it.

type ExactEra (inEra ∷ Type → Type) era = ProtVerInBounds era (ProtVerLow (inEra (EraCrypto era))) (ProtVerHigh (inEra (EraCrypto era)))

Restrict an era to the specific era through the protocol version. This is equivalent to (inEra (Crypto era) ~ era)

type family ProtVerAtLeast era (l ∷ Nat) where ...

Requirement for the era's highest protocol version to be higher or equal to the supplied value

Equations

ProtVerAtLeast era l = ProtVerIsInBounds "at least" era l (l <=? ProtVerHigh era)

type family ProtVerAtMost era (h ∷ Nat) where ...

Requirement for the era's lowest protocol version to be lower or equal to the supplied value

Equations

ProtVerAtMost era h = ProtVerIsInBounds "at most" era h (ProtVerLow era <=? h)

type ProtVerInBounds era (l ∷ Nat) (h ∷ Nat) = (ProtVerAtLeast era l, ProtVerAtMost era h)

Restrict a lower and upper bounds of the protocol version for the particular era

atLeastEra ∷ ∀ (eraName ∷ Type → Type) era. AtLeastEra eraName era ⇒ ()

Enforce era to be at least the specified era at the type level. In other words compiler will produce type error when applied to eras prior to the specified era. This function should be used in order to avoid redundant constraints warning.

For example these will type check

>>> atLeastEra @BabbageEra @(ConwayEra StandardCrypto)
>>> atLeastEra @BabbageEra @(BabbageEra StandardCrypto)

However this will result in a type error

>>> atLeastEra @BabbageEra @(AlonzoEra StandardCrypto)

atMostEra ∷ ∀ (eraName ∷ Type → Type) era. AtMostEra eraName era ⇒ ()

Enforce era to be at most the specified era at the type level. In other words compiler will produce type error when applied to eras prior to the specified era. This function should be used in order to avoid redundant constraints warning.

For example these will type check

>>> atMostEra @BabbageEra @(ShelleyEra StandardCrypto)
>>> atMostEra @AlonzoEra @(MaryEra StandardCrypto)

However this will result in a type error

>>> atMostEra @BabbageEra @(ConwayEra StandardCrypto)