Cardano.Ledger.Api.Governance

Documentation

class (EraPParams era, Eq (GovState era), Show (GovState era), NoThunks (GovState era), NFData (GovState era), EncCBOR (GovState era), DecCBOR (GovState era), DecShareCBOR (GovState era), Share (GovState era) ~ (Interns (Credential 'Staking), Interns (KeyHash 'StakePool), Interns (Credential 'DRepRole), Interns (Credential 'HotCommitteeRole)), ToCBOR (GovState era), FromCBOR (GovState era), Default (GovState era), ToJSON (GovState era)) ⇒ EraGov era

Minimal complete definition

curPParamsGovStateL, prevPParamsGovStateL, futurePParamsGovStateL, obligationGovState

Associated Types

type GovState era = (r ∷ Type) | r → era

emptyGovState ∷ EraGov era ⇒ GovState era

Construct empty governance state

curPParamsGovStateL ∷ EraGov era ⇒ Lens' (GovState era) (PParams era)

Lens for accessing current protocol parameters

prevPParamsGovStateL ∷ EraGov era ⇒ Lens' (GovState era) (PParams era)

Lens for accessing the previous protocol parameters

futurePParamsGovStateL ∷ EraGov era ⇒ Lens' (GovState era) (FuturePParams era)

Lens for accessing the future protocol parameters.

This lens will produce DefinitePParamsUpdate whenever we are absolutely sure that the new PParams will be updated. Which means there will be no chance of a DefinitePParamsUpdate value until we are past the point of no return, which is 2 stability windows before the end of the epoch. This lens is mostly intended for ledger usage and nextEpochUpdatedPParams should be used instead whenever definite results are desired.

Shelley

data ShelleyGovState era

Constructors

ShelleyGovState

Fields

sgsCurProposals ∷ !(ProposedPPUpdates era)

sgsFutureProposals ∷ !(ProposedPPUpdates era)

sgsCurPParams ∷ !(PParams era)

sgsPrevPParams ∷ !(PParams era)

sgsFuturePParams ∷ !(FuturePParams era)

Prediction of parameter changes that might happen on the epoch boundary.

newtype ProposedPPUpdates era

Update operation for protocol parameters structure PParams

Constructors

ProposedPPUpdates (Map (KeyHash 'Genesis) (PParamsUpdate era))

emptyPPPUpdates ∷ ProposedPPUpdates era

Conway

Governance Procedures

data VotingProcedure era

Constructors

VotingProcedure

Fields

vProcVote ∷ !Vote

vProcAnchor ∷ !(StrictMaybe Anchor)

newtype VotingProcedures era

Constructors

VotingProcedures

Fields

unVotingProcedures ∷ Map Voter (Map GovActionId (VotingProcedure era))

data ProposalProcedure era

Constructors

ProposalProcedure

Fields

pProcDeposit ∷ !Coin

pProcReturnAddr ∷ !RewardAccount

pProcGovAction ∷ !(GovAction era)

pProcAnchor ∷ !Anchor

Constitution

data Constitution era

Constructors

Constitution

Fields

constitutionAnchor ∷ !Anchor

constitutionScript ∷ !(StrictMaybe ScriptHash)

constitutionAnchorL ∷ ∀ era f. Functor f ⇒ (Anchor → f Anchor) → Constitution era → f (Constitution era)

constitutionScriptL ∷ ∀ era f. Functor f ⇒ (StrictMaybe ScriptHash → f (StrictMaybe ScriptHash)) → Constitution era → f (Constitution era)

Governance State

data ConwayGovState era

Conway governance state

Constructors

ConwayGovState

Fields

cgsProposals ∷ !(Proposals era)

cgsCommittee ∷ !(StrictMaybe (Committee era))

cgsConstitution ∷ !(Constitution era)

cgsCurPParams ∷ !(PParams era)

cgsPrevPParams ∷ !(PParams era)

cgsFuturePParams ∷ !(FuturePParams era)

cgsDRepPulsingState ∷ !(DRepPulsingState era)

The cgsDRepPulsingState field is a pulser that incrementally computes the stake distribution of the DReps over the Epoch following the close of voting at end of the previous Epoch. It assembles this with some of its other internal components into a (RatifyEnv era) when it completes, and then calls the RATIFY rule and eventually returns the updated RatifyState. The pulser is created at the Epoch boundary, but does no work until it is pulsed in the NEWEPOCH rule, whenever the system is NOT at the epoch boundary.

cgsProposalsL ∷ ∀ era f. Functor f ⇒ (Proposals era → f (Proposals era)) → ConwayGovState era → f (ConwayGovState era)

data RatifyState era

RatifyState stores information about what will happen to the active governance actions at the next epoch boundary.

Constructors

RatifyState

Fields

rsEnactState ∷ !(EnactState era)

This is the currently active EnactState. It contains all the changes that were applied to it at the last epoch boundary by all the proposals that were enacted.

rsEnacted ∷ !(Seq (GovActionState era))

Governance actions that are going to be enacted at the next epoch boundary.

rsExpired ∷ !(Set GovActionId)

Governance actions that are going to be removed at the next epoch boundary, either due to expiring or because they would become invalid after another governance action gets enacted or expired before it

rsDelayed ∷ !Bool

This flag is set to true if one of the proposals that was ratified at the last epoch boundary was a delaying action. This means that no other proposals will be ratified this epoch and each active proposal that has not become invalid will have its expiry date extended by one epoch.

This flag is reset at each epoch boundary before the RATIFY rule gets called, but it might immediately be set to True again after the RATIFY rule has finished execution.

data EnactState era

Constructors

EnactState

Fields

ensCommittee ∷ !(StrictMaybe (Committee era))

Constitutional Committee

ensConstitution ∷ !(Constitution era)

Constitution

ensCurPParams ∷ !(PParams era)

ensPrevPParams ∷ !(PParams era)

ensTreasury ∷ !Coin

ensWithdrawals ∷ !(Map (Credential 'Staking) Coin)

ensPrevGovActionIds ∷ !(GovRelation StrictMaybe)

Last enacted GovAction Ids

data Voter

Constructors

CommitteeVoter !(Credential 'HotCommitteeRole)

DRepVoter !(Credential 'DRepRole)

StakePoolVoter !(KeyHash 'StakePool)

data Vote

Constructors

VoteNo

VoteYes

Abstain

Governance Action

data GovAction era

Note that the previous governance action id is only optional for the very first governance action of the same purpose.

Constructors

ParameterChange

Fields

!(StrictMaybe (GovPurposeId 'PParamUpdatePurpose))

Previous governance action id of ParameterChange type, which corresponds to PParamUpdatePurpose.

!(PParamsUpdate era)

Proposed changes to PParams

!(StrictMaybe ScriptHash)

Policy hash protection

HardForkInitiation

Fields

!(StrictMaybe (GovPurposeId 'HardForkPurpose))

Previous governance action id of HardForkInitiation type, which corresponds to HardForkPurpose

!ProtVer

Proposed new protocol version

TreasuryWithdrawals

Fields

!(Map RewardAccount Coin)

Proposed treasury withdrawals

!(StrictMaybe ScriptHash)

Policy hash protection

NoConfidence

Fields

!(StrictMaybe (GovPurposeId 'CommitteePurpose))

Previous governance action id of NoConfidence or UpdateCommittee type, which corresponds to CommitteePurpose

UpdateCommittee

Fields

!(StrictMaybe (GovPurposeId 'CommitteePurpose))

Previous governance action id of UpdateCommittee or NoConfidence type, which corresponds to CommitteePurpose

!(Set (Credential 'ColdCommitteeRole))

Constitutional Committe members to be removed

!(Map (Credential 'ColdCommitteeRole) EpochNo)

Constitutional committee members to be added

!UnitInterval

New Threshold

NewConstitution

Fields

!(StrictMaybe (GovPurposeId 'ConstitutionPurpose))

Previous governance action id of NewConstitution type, which corresponds to ConstitutionPurpose

!(Constitution era)

InfoAction

data GovActionId

Constructors

GovActionId

Fields

gaidTxId ∷ !TxId

gaidGovActionIx ∷ !GovActionIx

newtype GovActionIx

Constructors

GovActionIx

Fields

unGovActionIx ∷ Word16

data GovActionState era

Constructors

GovActionState

Fields

gasId ∷ !GovActionId

gasCommitteeVotes ∷ !(Map (Credential 'HotCommitteeRole) Vote)

gasDRepVotes ∷ !(Map (Credential 'DRepRole) Vote)

gasStakePoolVotes ∷ !(Map (KeyHash 'StakePool) Vote)

gasProposalProcedure ∷ !(ProposalProcedure era)

gasProposedIn ∷ !EpochNo

gasExpiresAfter ∷ !EpochNo

data GovActionPurpose

Constructors

PParamUpdatePurpose

HardForkPurpose

CommitteePurpose

ConstitutionPurpose

data GovRelation (f ∷ Type → Type)

Abstract data type for representing relationship of governance action with the same purpose

Constructors

GovRelation

Fields

grPParamUpdate ∷ !(f (GovPurposeId 'PParamUpdatePurpose))

grHardFork ∷ !(f (GovPurposeId 'HardForkPurpose))

grCommittee ∷ !(f (GovPurposeId 'CommitteePurpose))

grConstitution ∷ !(f (GovPurposeId 'ConstitutionPurpose))

hoistGovRelation ∷ (∀ a. f a → g a) → GovRelation f → GovRelation g

withGovActionParent

Arguments

∷ GovActionState era

→ a
The result to be used for governance actions that can't have a parent

→ (∀ (p ∷ GovActionPurpose). (∀ (f1 ∷ Type → Type) (f2 ∷ Type → Type). Functor f2 ⇒ (f1 (GovPurposeId p) → f2 (f1 (GovPurposeId p))) → GovRelation f1 → f2 (GovRelation f1)) → StrictMaybe (GovPurposeId p) → GovPurposeId p → a)
Function that will be applied to a lens and a parent

→ a

Apply a function to a GovAction that can have a parent.

newtype GovPurposeId (p ∷ GovActionPurpose)

Constructors

GovPurposeId

Fields

unGovPurposeId ∷ GovActionId

govActionIdToText ∷ GovActionId → Text

Anchor

data Anchor

Constructors

Anchor

Fields

anchorUrl ∷ !Url

anchorDataHash ∷ !(SafeHash AnchorData)

newtype AnchorData

Constructors

AnchorData ByteString

hashAnchorData ∷ AnchorData → SafeHash AnchorData

Hash AnchorData