Cardano.Ledger.Api.PParams

PParams

data PParams era

Protocol parameters

emptyPParams ∷ EraPParams era ⇒ PParams era

type family UpgradePParams (f ∷ Type → Type) era

upgradePParams ∷ (EraPParams era, EraPParams (PreviousEra era)) ⇒ UpgradePParams Identity era → PParams (PreviousEra era) → PParams era

type family DowngradePParams (f ∷ Type → Type) era

downgradePParams ∷ (EraPParams era, EraPParams (PreviousEra era)) ⇒ DowngradePParams Identity era → PParams era → PParams (PreviousEra era)

PParamsUpdate

data PParamsUpdate era

The type of updates to Protocol parameters

emptyPParamsUpdate ∷ EraPParams era ⇒ PParamsUpdate era

upgradePParamsUpdate ∷ (EraPParams era, EraPParams (PreviousEra era)) ⇒ UpgradePParams StrictMaybe era → PParamsUpdate (PreviousEra era) → PParamsUpdate era

downgradePParamsUpdate ∷ (EraPParams era, EraPParams (PreviousEra era)) ⇒ DowngradePParams StrictMaybe era → PParamsUpdate era → PParamsUpdate (PreviousEra era)

Shelley params

Protocol parameters introduced in Shelley era

MinFeeA

Min fee factor

ppMinFeeAL ∷ EraPParams era ⇒ Lens' (PParams era) Coin

The linear factor for the minimum fee calculation

ppuMinFeeAL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

The linear factor for the minimum fee calculation

MinFeeB

Min fee constant

ppMinFeeBL ∷ EraPParams era ⇒ Lens' (PParams era) Coin

The constant factor for the minimum fee calculation

ppuMinFeeBL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

The constant factor for the minimum fee calculation

MaxBBSize

Max block body size

ppMaxBBSizeL ∷ EraPParams era ⇒ Lens' (PParams era) Word32

Maximal block body size

ppuMaxBBSizeL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Word32)

Maximal block body size

AaxBHSize

Max block header size

ppMaxBHSizeL ∷ EraPParams era ⇒ Lens' (PParams era) Word16

Maximal block header size

ppuMaxBHSizeL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Word16)

Maximal block header size

PoolDeposit

Stake pool deposit

ppPoolDepositL ∷ EraPParams era ⇒ Lens' (PParams era) Coin

The amount of a pool registration deposit

ppuPoolDepositL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

The amount of a pool registration deposit

EMax

Epoch bound on pool retirement

ppEMaxL ∷ EraPParams era ⇒ Lens' (PParams era) EpochInterval

epoch bound on pool retirement

ppuEMaxL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe EpochInterval)

epoch bound on pool retirement

NOpt

Desired number of pools

ppNOptL ∷ EraPParams era ⇒ Lens' (PParams era) Natural

Desired number of pools

ppuNOptL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Natural)

Desired number of pools

A0

Pool influence

ppA0L ∷ EraPParams era ⇒ Lens' (PParams era) NonNegativeInterval

Pool influence

ppuA0L ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe NonNegativeInterval)

Pool influence

Tau

Treasury expansion

ppTauL ∷ EraPParams era ⇒ Lens' (PParams era) UnitInterval

Treasury expansion

ppuTauL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe UnitInterval)

Treasury expansion

Rho

Monetary expansion

ppRhoL ∷ EraPParams era ⇒ Lens' (PParams era) UnitInterval

Monetary expansion

ppuRhoL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe UnitInterval)

Monetary expansion

D

Decentralization paramteter

Note - Removed in Babbage

ppDL ∷ (EraPParams era, ProtVerAtMost era 6) ⇒ Lens' (PParams era) UnitInterval

Decentralization parameter

ppDG ∷ EraPParams era ⇒ SimpleGetter (PParams era) UnitInterval

Decentralization parameter getter

ppuDL ∷ (EraPParams era, ProtVerAtMost era 6) ⇒ Lens' (PParamsUpdate era) (StrictMaybe UnitInterval)

Decentralization parameter

ExtraEntropy

Extra entropy

Note - Removed in Babbage

ppExtraEntropyL ∷ (EraPParams era, ProtVerAtMost era 6) ⇒ Lens' (PParams era) Nonce

Extra entropy

ppuExtraEntropyL ∷ (EraPParams era, ProtVerAtMost era 6) ⇒ Lens' (PParamsUpdate era) (StrictMaybe Nonce)

Extra entropy

ProtocolVersion

Protocol version

ppProtocolVersionL ∷ EraPParams era ⇒ Lens' (PParams era) ProtVer

ppuProtocolVersionL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe ProtVer)

PParamsUpdate Protocol version

MinUTxOValue

Minimum allowed value of a new TxOut

ppuMinUTxOValueL ∷ (EraPParams era, ProtVerAtMost era 4) ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

Minimum UTxO value

ppMinUTxOValueL ∷ (EraPParams era, ProtVerAtMost era 4) ⇒ Lens' (PParams era) Coin

Minimum UTxO value

MinPoolCast

Miminum allowed stake pool cost

ppMinPoolCostL ∷ EraPParams era ⇒ Lens' (PParams era) Coin

Minimum Stake Pool Cost

ppuMinPoolCostL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

Minimum Stake Pool Cost

KeyDeposit

ppKeyDepositL ∷ EraPParams era ⇒ Lens' (PParams era) Coin

The amount of a key registration deposit

ppuKeyDepositL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

The amount of a key registration deposit

MaxTxSize

ppMaxTxSizeL ∷ EraPParams era ⇒ Lens' (PParams era) Word32

Maximal transaction size

ppuMaxTxSizeL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Word32)

Maximal transaction size

Alonzo params

Protocol parameters introduced in Alonzo era

data UpgradeAlonzoPParams (f ∷ Type → Type)

Parameters that were added in Alonzo

Constructors

UpgradeAlonzoPParams

Fields

uappCoinsPerUTxOWord ∷ !(HKD f CoinPerWord)

uappCostModels ∷ !(HKD f CostModels)

uappPrices ∷ !(HKD f Prices)

uappMaxTxExUnits ∷ !(HKD f ExUnits)

uappMaxBlockExUnits ∷ !(HKD f ExUnits)

uappMaxValSize ∷ !(HKD f Natural)

uappCollateralPercentage ∷ !(HKD f Natural)

uappMaxCollateralInputs ∷ !(HKD f Natural)

newtype DowngradeAlonzoPParams (f ∷ Type → Type)

Parameters that were removed in Alonzo

Constructors

DowngradeAlonzoPParams

Fields

dappMinUTxOValue ∷ HKD f Coin

data LangDepView

Constructors

LangDepView

Fields

tag ∷ ByteString

params ∷ ByteString

getLanguageView ∷ AlonzoEraPParams era ⇒ PParams era → Language → LangDepView

CostModels

Plutus CostModels

To convert resource primitives into the more abstract ExUnits during script execution a cost model needs to be supplied to the interpreter. The cost models required for this purpose are recorded in the CostModels protocol parameter.

ppCostModelsL ∷ AlonzoEraPParams era ⇒ Lens' (PParams era) CostModels

ppuCostModelsL ∷ AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe CostModels)

Prices

The calculation of the actual cost, in Ada, of running a script that takes ExUnits resources to run, is done by a formula in the ledger rules, which uses the Prices parameter

ppPricesL ∷ AlonzoEraPParams era ⇒ Lens' (PParams era) Prices

ppuPricesL ∷ AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Prices)

MaxTxExUnits

Limit the total per-transaction resource use for phase-2 scripts.

ppMaxTxExUnitsL ∷ AlonzoEraPParams era ⇒ Lens' (PParams era) ExUnits

ppuMaxTxExUnitsL ∷ AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe ExUnits)

MaxBlockExUnits

Limit the total per-transaction and per-block resource use for phase-2 scripts.

ppMaxBlockExUnitsL ∷ AlonzoEraPParams era ⇒ Lens' (PParams era) ExUnits

ppuMaxBlockExUnitsL ∷ AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe ExUnits)

MaxValSize

The new parameter maxValSize replaces the constant maxValSize used Mary era to limit the size of the Value part of an output in a serialised transaction.

ppMaxValSizeL ∷ AlonzoEraPParams era ⇒ Lens' (PParams era) Natural

ppuMaxValSizeL ∷ AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Natural)

CoinsPerUTxOWord

In Alonzo, the protocol parameter minUTxOValue is deprecated, and replaced by coinsPerUTxOWord. This specifies directly the deposit required for storing bytes of data on the ledger in the form of UTxO entries.

newtype CoinPerWord

Constructors

CoinPerWord

Fields

unCoinPerWord ∷ Coin

ppCoinsPerUTxOWordL ∷ (AlonzoEraPParams era, ExactEra AlonzoEra era) ⇒ Lens' (PParams era) CoinPerWord

ppuCoinsPerUTxOWordL ∷ (AlonzoEraPParams era, ExactEra AlonzoEra era) ⇒ Lens' (PParamsUpdate era) (StrictMaybe CoinPerWord)

CollateralPercentage

The parameter collateralPercent is used to specify the percentage of the total transaction fee its collateral must (at minimum) cover.

ppCollateralPercentageL ∷ AlonzoEraPParams era ⇒ Lens' (PParams era) Natural

ppuCollateralPercentageL ∷ AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Natural)

MaxCollateralInputs

The parameter maxCollateralInputs is used to limit, additionally, the total number of collateral inputs, and thus the total number of additional signatures that must be checked during validation.

ppMaxCollateralInputsL ∷ AlonzoEraPParams era ⇒ Lens' (PParams era) Natural

ppuMaxCollateralInputsL ∷ AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Natural)

Babbage params

Protocol parameters introduced in Babbage era

data DowngradeBabbagePParams (f ∷ Type → Type)

Constructors

DowngradeBabbagePParams

Fields

dbppD ∷ !(HKD f UnitInterval)

dbppExtraEntropy ∷ !(HKD f Nonce)

CoinsPerUTxOByte

Cost in the amount of lovelace ber byte.

Note - This parameter is a replacement for coinsPerUTxOWord

newtype CoinPerByte

Constructors

CoinPerByte

Fields

unCoinPerByte ∷ Coin

coinsPerUTxOWordToCoinsPerUTxOByte ∷ CoinPerWord → CoinPerByte

A word is 8 bytes, so convert from coinsPerUTxOWord to coinsPerUTxOByte, rounding down.

coinsPerUTxOByteToCoinsPerUTxOWord ∷ CoinPerByte → CoinPerWord

A word is 8 bytes, so convert from coinsPerUTxOByte to coinsPerUTxOWord.

ppCoinsPerUTxOByteL ∷ BabbageEraPParams era ⇒ Lens' (PParams era) CoinPerByte

ppuCoinsPerUTxOByteL ∷ BabbageEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe CoinPerByte)

Type classes

class (Era era, Eq (PParamsHKD Identity era), Ord (PParamsHKD Identity era), Show (PParamsHKD Identity era), NFData (PParamsHKD Identity era), EncCBOR (PParamsHKD Identity era), DecCBOR (PParamsHKD Identity era), ToCBOR (PParamsHKD Identity era), FromCBOR (PParamsHKD Identity era), NoThunks (PParamsHKD Identity era), ToJSON (PParamsHKD Identity era), FromJSON (PParamsHKD Identity era), Eq (PParamsHKD StrictMaybe era), Ord (PParamsHKD StrictMaybe era), Show (PParamsHKD StrictMaybe era), NFData (PParamsHKD StrictMaybe era), EncCBOR (PParamsHKD StrictMaybe era), DecCBOR (PParamsHKD StrictMaybe era), ToCBOR (PParamsHKD StrictMaybe era), FromCBOR (PParamsHKD StrictMaybe era), NoThunks (PParamsHKD StrictMaybe era), ToJSON (PParamsHKD StrictMaybe era)) ⇒ EraPParams era

Minimal complete definition

emptyPParamsIdentity, emptyPParamsStrictMaybe, upgradePParamsHKD, downgradePParamsHKD, hkdMinFeeAL, hkdMinFeeBL, hkdMaxBBSizeL, hkdMaxTxSizeL, hkdMaxBHSizeL, hkdKeyDepositL, hkdPoolDepositL, hkdEMaxL, hkdNOptL, hkdA0L, hkdRhoL, hkdTauL, hkdDL, hkdExtraEntropyL, hkdProtocolVersionL, hkdMinUTxOValueL, hkdMinPoolCostL

class EraPParams era ⇒ AlonzoEraPParams era

Minimal complete definition

hkdCoinsPerUTxOWordL, hkdCostModelsL, hkdPricesL, hkdMaxTxExUnitsL, hkdMaxBlockExUnitsL, hkdMaxValSizeL, hkdCollateralPercentageL, hkdMaxCollateralInputsL

class AlonzoEraPParams era ⇒ BabbageEraPParams era

Minimal complete definition

hkdCoinsPerUTxOByteL