Cardano.Ledger.Api.Scripts

Documentation

module Cardano.Ledger.Api.Scripts.Data

class (Era era, Show (Script era), Eq (Script era), ToCBOR (Script era), EncCBOR (Script era), DecCBOR (Annotator (Script era)), NoThunks (Script era), SafeToHash (Script era)) ⇒ EraScript era where

Typeclass for script data types. Allows for script validation and hashing. You must understand the role of SafeToHash and scriptPrefixTag to make new instances. scriptPrefixTag is a magic number representing the tag of the script language. For each new script language defined, a new tag is chosen and the tag is included in the script hash for a script. The safeToHash constraint ensures that Scripts are never reserialised.

Minimal complete definition

scriptPrefixTag, phaseScript

Associated Types

type Script era = (r ∷ Type) | r → era

Scripts which may lock transaction outputs in this era

Methods

scriptPrefixTag ∷ Script era → ByteString

hashScript ∷ Script era → ScriptHash (EraCrypto era)

phaseScript ∷ ∀ (phase ∷ Phase). PhaseRep phase → Script era → Maybe (PhasedScript phase era)

isNativeScript ∷ Script era → Bool

data ScriptHash c

data CostModels

For a known version of Plutus, attempting to construct a cost model with too few parameters (depending on the version) will result in an error. CostModelApplyError exists to collect these errors in the CostModels type. The CostModels type itself needs to be flexible enough to accept any map of Word8 to '[Integer]', so that cost models can be placed in the protocol parameters ahead of changes to the Plutus evaluation context. In this way, serializing a cost model, updating software, and deserializing can result in errors going away.

Additionally, CostModels needs to be able to store cost models for future version of Plutus, which we cannot yet even validate. These are stored in invalidCostModels.

Constructors

CostModels

Fields

costModelsValid ∷ !(Map Language CostModel)

costModelsErrors ∷ !(Map Language CostModelError)

costModelsUnknown ∷ !(Map Word8 [Integer])

data ValidityInterval

ValidityInterval is a half open interval. Closed on the bottom, open on the top. A SNothing on the bottom is negative infinity, and a SNothing on the top is positive infinity

Constructors

ValidityInterval

Fields

invalidBefore ∷ !(StrictMaybe SlotNo)

invalidHereafter ∷ !(StrictMaybe SlotNo)