Cardano.Ledger.Api.State.Query

GetFilteredDelegationsAndRewardAccounts

filterStakePoolDelegsAndRewards ∷ UMap c → Set (Credential 'Staking c) → (Map (Credential 'Staking c) (KeyHash 'StakePool c), Map (Credential 'Staking c) Coin)

Filter out stake pool delegations and rewards for a set of stake credentials

queryStakePoolDelegsAndRewards ∷ NewEpochState era → Set (Credential 'Staking (EraCrypto era)) → (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)), Map (Credential 'Staking (EraCrypto era)) Coin)

Uses filterStakePoolDelegsAndRewards to get the same information from the NewEpochState

Implementation for GetFilteredDelegationsAndRewardAccounts query.

GetGovState

queryGovState ∷ NewEpochState era → GovState era

This query returns all of the state related to governance

GetConstitution

queryConstitution ∷ ConwayEraGov era ⇒ NewEpochState era → Constitution era

GetConstitutionHash

queryConstitutionHash ∷ ConwayEraGov era ⇒ NewEpochState era → SafeHash (EraCrypto era) AnchorData

GetDRepState

queryDRepState

Arguments

∷ NewEpochState era

→ Set (Credential 'DRepRole (EraCrypto era))
Specify a set of DRep credentials whose state should be returned. When this set is empty, states for all of the DReps will be returned.

→ Map (Credential 'DRepRole (EraCrypto era)) (DRepState (EraCrypto era))

Query DRep state.

GetDRepStakeDistr

queryDRepStakeDistr

Arguments

∷ ConwayEraGov era

⇒ NewEpochState era

→ Set (DRep (EraCrypto era))
Specify DRep Ids whose stake distribution should be returned. When this set is empty, distributions for all of the DReps will be returned.

→ Map (DRep (EraCrypto era)) Coin

Query DRep stake distribution. Note that this can be an expensive query because there is a chance that current distribution has not been fully computed yet.

GetSPOStakeDistr

querySPOStakeDistr

Arguments

∷ ConwayEraGov era

⇒ NewEpochState era

→ Set (KeyHash 'StakePool (EraCrypto era))

→ Map (KeyHash 'StakePool (EraCrypto era)) Coin
Specify pool key hashes whose stake distribution should be returned. When this set is empty, distributions for all of the pools will be returned.

Query pool stake distribution.

GetCommitteeState

queryCommitteeState ∷ NewEpochState era → CommitteeState era

Deprecated: In favor of queryCommitteeMembersState

Query committee members

GetCommitteeMembersState

queryCommitteeMembersState

Arguments

∷ ∀ era. ConwayEraGov era

⇒ Set (Credential 'ColdCommitteeRole (EraCrypto era))
filter by cold credentials (don't filter when empty)

→ Set (Credential 'HotCommitteeRole (EraCrypto era))
filter by hot credentials (don't filter when empty)

→ Set MemberStatus
filter by status (don't filter when empty) (useful, for discovering, for example, only active members)

→ NewEpochState era

→ CommitteeMembersState (EraCrypto era)

Query committee members. Whenever the system is in No Confidence mode this query will return Nothing.

GetAccountState

queryAccountState ∷ NewEpochState era → AccountState

data CommitteeMemberState c

Constructors

CommitteeMemberState

Fields

cmsHotCredAuthStatus ∷ !(HotCredAuthStatus c)

cmsStatus ∷ !MemberStatus

cmsExpiration ∷ !(Maybe EpochNo)

Absolute epoch number when the member expires

cmsNextEpochChange ∷ !NextEpochChange

Changes to the member at the next epoch

data CommitteeMembersState c

Constructors

CommitteeMembersState

Fields

csCommittee ∷ !(Map (Credential 'ColdCommitteeRole c) (CommitteeMemberState c))

csThreshold ∷ !(Maybe UnitInterval)

csEpochNo ∷ !EpochNo

Current epoch number. This is necessary to interpret committee member states

data HotCredAuthStatus c

Constructors

MemberAuthorized (Credential 'HotCommitteeRole c)

MemberNotAuthorized
Member enacted, but no hot credential for voting has been registered

MemberResigned (Maybe (Anchor c))

data MemberStatus

Constructors

Active

Expired

Unrecognized
This can happen when a hot credential for an unknown cold credential exists. Such Committee member will be either removed from the state at the next epoch boundary or enacted as a new member.

data NextEpochChange

Constructors

ToBeEnacted

ToBeRemoved
Member will be removed

NoChangeExpected

ToBeExpired

TermAdjusted EpochNo

GetCurrentPParams

queryCurrentPParams ∷ EraGov era ⇒ NewEpochState era → PParams era

This is a simple lookup into the state for the values of current protocol parameters. These values can change on the epoch boundary. Use queryFuturePParams to see if we are aware of any upcoming changes.

GetFuturePParams

queryFuturePParams ∷ EraGov era ⇒ NewEpochState era → Maybe (PParams era)

This query will return values for protocol parameters that are likely to be adopted at the next epoch boundary. It is only when we passed 2 stability windows before the end of the epoch that users can rely on this query to produce stable results.

GetProposals

queryProposals

Arguments

∷ ConwayEraGov era

⇒ NewEpochState era

→ Set (GovActionId (EraCrypto era))
Specify a set of Governance Action IDs to filter the proposals. When this set is empty, all the proposals considered for ratification will be returned.

→ Seq (GovActionState era)

Query proposals that are considered for ratification.

GetRatifyState

queryRatifyState ∷ ConwayEraGov era ⇒ NewEpochState era → RatifyState era

Query ratification state.

For testing

getNextEpochCommitteeMembers ∷ ConwayEraGov era ⇒ NewEpochState era → Map (Credential 'ColdCommitteeRole (EraCrypto era)) EpochNo