Cardano.Ledger.Api.State.Query

GetFilteredDelegationsAndRewardAccounts

filterStakePoolDelegsAndRewards :: UMap c -> Set (Credential 'Staking c) -> (Map (Credential 'Staking c) (KeyHash 'StakePool c), Map (Credential 'Staking c) Coin)

Filter out stake pool delegations and rewards for a set of stake credentials

queryStakePoolDelegsAndRewards :: NewEpochState era -> Set (Credential 'Staking (EraCrypto era)) -> (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)), Map (Credential 'Staking (EraCrypto era)) Coin)

Uses filterStakePoolDelegsAndRewards to get the same information from the NewEpochState

Implementation for GetFilteredDelegationsAndRewardAccounts query.

GetGovState

queryGovState :: NewEpochState era -> GovState era

This query returns all of the state related to governance

GetConstitution

queryConstitution :: EraGov era => NewEpochState era -> Maybe (Constitution era)

GetConstitutionHash

queryConstitutionHash :: EraGov era => NewEpochState era -> Maybe (SafeHash (EraCrypto era) AnchorData)

GetDRepState

queryDRepState

Arguments

:: NewEpochState era

-> Set (Credential 'DRepRole (EraCrypto era))
Specify a set of DRep credentials whose state should be returned. When this set is empty, states for all of the DReps will be returned.

-> Map (Credential 'DRepRole (EraCrypto era)) (DRepState (EraCrypto era))

Query DRep state.

GetDRepStakeDistr

queryDRepStakeDistr

Arguments

:: EraGov era

=> NewEpochState era

-> Set (DRep (EraCrypto era))
Specify DRep Ids whose stake distribution should be returned. When this set is empty, distributions for all of the DReps will be returned.

-> Map (DRep (EraCrypto era)) Coin

Query DRep stake distribution. Note that this can be an expensive query because there is a chance that current distribution has not been fully computed yet.

GetCommitteeState

queryCommitteeState :: NewEpochState era -> CommitteeState era

Deprecated: In favor of queryCommitteeMembersState

Query committee members

GetCommitteeMembersState

queryCommitteeMembersState

Arguments

:: forall era. EraGov era

=> Set (Credential 'ColdCommitteeRole (EraCrypto era))
filter by cold credentials (don't filter when empty)

-> Set (Credential 'HotCommitteeRole (EraCrypto era))
filter by hot credentials (don't filter when empty)

-> Set MemberStatus
filter by status (don't filter when empty) (useful, for discovering, for example, only active members)

-> NewEpochState era

-> Maybe (CommitteeMembersState (EraCrypto era))

Query committee members. Whenever the system is in No Confidence mode this query will return Nothing.

data CommitteeMemberState c

Constructors

CommitteeMemberState

Fields

cmsHotCredAuthStatus :: !(HotCredAuthStatus c)

cmsStatus :: !MemberStatus

cmsExpiration :: !(Maybe EpochNo)

Absolute epoch number when the member expires

cmsNextEpochChange :: !NextEpochChange

Changes to the member at the next epoch

data CommitteeMembersState c

Constructors

CommitteeMembersState

Fields

csCommittee :: !(Map (Credential 'ColdCommitteeRole c) (CommitteeMemberState c))

csQuorum :: !UnitInterval

csEpochNo :: !EpochNo

Current epoch number. This is necessary to interpret committee member states

data HotCredAuthStatus c

Constructors

MemberAuthorized (Credential 'HotCommitteeRole c)

MemberNotAuthorized
Member enacted, but no hot credential for voting has been registered

MemberResigned

data MemberStatus

Constructors

Active

Expired

Unrecognized
This can happen when a hot credential for an unknown cold credential exists. Such Committee member will be either removed from the state at the next epoch boundary or enacted as a new member.

data NextEpochChange

Constructors

ToBeEnacted

ToBeRemoved
Member will be removed

NoChangeExpected