Cardano.Ledger.Api.State.Query

GetFilteredDelegationsAndRewardAccounts

filterStakePoolDelegsAndRewards ∷ UMap c → Set (Credential 'Staking c) → (Map (Credential 'Staking c) (KeyHash 'StakePool c), Map (Credential 'Staking c) Coin)

Filter out stake pool delegations and rewards for a set of stake credentials

queryStakePoolDelegsAndRewards ∷ NewEpochState era → Set (Credential 'Staking (EraCrypto era)) → (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)), Map (Credential 'Staking (EraCrypto era)) Coin)

Uses filterStakePoolDelegsAndRewards to get the same information from the NewEpochState

Implementation for GetFilteredDelegationsAndRewardAccounts query.

GetConstitutionHash

queryConstitutionHash ∷ EraGovernance era ⇒ NewEpochState era → Maybe (SafeHash (EraCrypto era) ByteString)