Cardano.Ledger.Api.Transition

Description

This module is used for defining initial configuration for all. It is also used in testing and benchmarking to initilize a chain in a particular era without going through the trouble of generating all the history for preceeding eras.

Documentation

class (EraTxOut era, EraGov era, ToJSON (TransitionConfig era), FromJSON (TransitionConfig era), Default (StashedAVVMAddresses era)) => EraTransition era

Minimal complete definition

mkTransitionConfig, tcPreviousEraConfigL, tcTranslationContextL

data family TransitionConfig era

mkLatestTransitionConfig :: Crypto c => ShelleyGenesis c -> AlonzoGenesis -> ConwayGenesis c -> TransitionConfig (LatestKnownEra c)

mkTransitionConfig :: EraTransition era => TranslationContext era -> TransitionConfig (PreviousEra era) -> TransitionConfig era

mkShelleyTransitionConfig :: ShelleyGenesis c -> TransitionConfig (ShelleyEra c)

tcShelleyGenesisL :: EraTransition era => Lens' (TransitionConfig era) (ShelleyGenesis (EraCrypto era))

tcPreviousEraConfigL :: EraTransition era => Lens' (TransitionConfig era) (TransitionConfig (PreviousEra era))

tcTranslationContextL :: EraTransition era => Lens' (TransitionConfig era) (TranslationContext era)

Genesis

data ShelleyGenesis c

Constructors

ShelleyGenesis

Fields

sgSystemStart :: !UTCTime

sgNetworkMagic :: !Word32

sgNetworkId :: !Network

sgActiveSlotsCoeff :: !PositiveUnitInterval

sgSecurityParam :: !Word64

sgEpochLength :: !EpochSize

sgSlotsPerKESPeriod :: !Word64

sgMaxKESEvolutions :: !Word64

sgSlotLength :: !NominalDiffTimeMicro

sgUpdateQuorum :: !Word64

sgMaxLovelaceSupply :: !Word64

sgProtocolParams :: !(PParams (ShelleyEra c))

sgGenDelegs :: !(Map (KeyHash 'Genesis c) (GenDelegPair c))

sgInitialFunds :: ListMap (Addr c) Coin

sgStaking :: ShelleyGenesisStaking c

newtype AlonzoGenesis

Constructors

AlonzoGenesisWrapper

Fields

unAlonzoGenesisWrapper :: UpgradeAlonzoPParams Identity

Bundled Patterns

pattern AlonzoGenesis :: CoinPerWord -> CostModels -> Prices -> ExUnits -> ExUnits -> Natural -> Natural -> Natural -> AlonzoGenesis

data ConwayGenesis c

Constructors

ConwayGenesis

Fields

cgUpgradePParams :: !(UpgradeConwayPParams Identity)

cgConstitution :: !(Constitution (ConwayEra c))

cgCommittee :: !(Committee (ConwayEra c))

Functions for Testing

tcInitialPParamsG :: EraTransition era => SimpleGetter (TransitionConfig era) (PParams era)

tcInitialFundsL :: (HasCallStack, EraTransition era) => Lens' (TransitionConfig era) (ListMap (Addr (EraCrypto era)) Coin)

tcInitialStakingL :: (HasCallStack, EraTransition era) => Lens' (TransitionConfig era) (ShelleyGenesisStaking (EraCrypto era))

createInitialState :: (EraTransition era, HasCallStack) => TransitionConfig era -> NewEpochState era

registerInitialFunds :: (EraTransition era, HasCallStack) => TransitionConfig era -> NewEpochState era -> NewEpochState era

registerInitialStaking :: (EraTransition era, HasCallStack) => TransitionConfig era -> NewEpochState era -> NewEpochState era