Cardano.Ledger.Api.Tx.Address

Address

data Addr c

An address for UTxO.

Contents of Addr data type are intentionally left as lazy, otherwise operating on compact form of an address will result in redundant work.

Constructors

Addr Network (PaymentCredential c) (StakeReference c)

AddrBootstrap (BootstrapAddress c)

newtype BootstrapAddress c

Constructors

BootstrapAddress

Fields

unBootstrapAddress ∷ Address

serialiseAddr ∷ Addr c → ByteString

Serialise an address to the external format.

decodeAddr ∷ (Crypto c, MonadFail m) ⇒ ByteString → m (Addr c)

decodeAddrEither ∷ Crypto c ⇒ ByteString → Either String (Addr c)

decodeAddrShort ∷ (Crypto c, MonadFail m) ⇒ ShortByteString → m (Addr c)

decodeAddrShortEither ∷ Crypto c ⇒ ShortByteString → Either String (Addr c)

getNetwork ∷ Addr c → Network

Reward Account

data RewardAcnt c

An account based address for rewards

Constructors

RewardAcnt

Fields

getRwdNetwork ∷ !Network

getRwdCred ∷ !(Credential 'Staking c)

mkRwdAcnt ∷ Network → Credential 'Staking c → RewardAcnt c

serialiseRewardAcnt ∷ RewardAcnt c → ByteString

Serialise a reward account to the external format.

deserialiseRewardAcnt ∷ Crypto c ⇒ ByteString → Maybe (RewardAcnt c)

Deserialise a reward account from the external format. This will fail if the input data is not in the right format (or if there is trailing data).