Cardano.Ledger.Api.Tx.AuxData

Documentation

class (Era era, Eq (TxAuxData era), EqRaw (TxAuxData era), Show (TxAuxData era), NoThunks (TxAuxData era), ToCBOR (TxAuxData era), EncCBOR (TxAuxData era), DecCBOR (Annotator (TxAuxData era)), HashAnnotated (TxAuxData era) EraIndependentTxAuxData (EraCrypto era)) ⇒ EraTxAuxData era

TxAuxData which may be attached to a transaction

Minimal complete definition

mkBasicTxAuxData, metadataTxAuxDataL, upgradeTxAuxData, hashTxAuxData, validateTxAuxData

Associated Types

type TxAuxData era = (r ∷ Type) | r → era

mkBasicTxAuxData ∷ EraTxAuxData era ⇒ TxAuxData era

metadataTxAuxDataL ∷ EraTxAuxData era ⇒ Lens' (TxAuxData era) (Map Word64 Metadatum)

upgradeTxAuxData ∷ EraTxAuxData era ⇒ TxAuxData (PreviousEra era) → TxAuxData era

Every era, except Shelley, must be able to upgrade a TxAuxData from a previous era.

Warning - Important to note that any memoized binary representation will not be preserved. If you need to retain underlying bytes you can use translateEraThroughCBOR

hashTxAuxData ∷ EraTxAuxData era ⇒ TxAuxData era → AuxiliaryDataHash (EraCrypto era)

validateTxAuxData ∷ EraTxAuxData era ⇒ ProtVer → TxAuxData era → Bool

Shelley

data ShelleyTxAuxData era where

Bundled Patterns

pattern ShelleyTxAuxData ∷ Era era ⇒ Map Word64 Metadatum → ShelleyTxAuxData era

data Metadatum

A generic metadatum type.

Constructors

Map ![(Metadatum, Metadatum)]

List ![Metadatum]

I !Integer

B !ByteString

S !Text

Allegra

class EraTxAuxData era ⇒ AllegraEraTxAuxData era

Minimal complete definition

timelockScriptsTxAuxDataL

timelockScriptsTxAuxDataL ∷ AllegraEraTxAuxData era ⇒ Lens' (TxAuxData era) (StrictSeq (Timelock era))

data AllegraTxAuxData era where

Bundled Patterns

pattern AllegraTxAuxData ∷ Era era ⇒ Map Word64 Metadatum → StrictSeq (Timelock era) → AllegraTxAuxData era

Alonzo

class AllegraEraTxAuxData era ⇒ AlonzoEraTxAuxData era

Minimal complete definition

plutusScriptsTxAuxDataL

plutusScriptsTxAuxDataL ∷ AlonzoEraTxAuxData era ⇒ Lens' (TxAuxData era) (Map Language (NonEmpty PlutusBinary))

data AlonzoTxAuxData era where

Bundled Patterns

pattern AlonzoTxAuxData ∷ (HasCallStack, AlonzoEraScript era) ⇒ Map Word64 Metadatum → StrictSeq (Timelock era) → Map Language (NonEmpty PlutusBinary) → AlonzoTxAuxData era
Construct auxiliary data. Make sure not to supply plutus script versions that are not supported in this era, because it will result in a runtime exception. Use mkAlonzoTxAuxData instead if you need runtime safety guarantees.

pattern AlonzoTxAuxData' ∷ Map Word64 Metadatum → StrictSeq (Timelock era) → Map Language (NonEmpty PlutusBinary) → AlonzoTxAuxData era

mkAlonzoTxAuxData ∷ (Foldable f, AlonzoEraScript era) ⇒ Map Word64 Metadatum → f (AlonzoScript era) → AlonzoTxAuxData era

Helper function that will construct Auxiliary data from Metadatum map and a list of scripts.

Note that the relative order of same type scripts will be preserved.

getAlonzoTxAuxDataScripts ∷ AlonzoEraScript era ⇒ AlonzoTxAuxData era → StrictSeq (AlonzoScript era)