Cardano.Ledger.Api.Tx.AuxData

Documentation

class (Era era, Eq (TxAuxData era), Show (TxAuxData era), NoThunks (TxAuxData era), ToCBOR (TxAuxData era), EncCBOR (TxAuxData era), DecCBOR (Annotator (TxAuxData era)), HashAnnotated (TxAuxData era) EraIndependentTxAuxData (EraCrypto era)) ⇒ EraTxAuxData era where

TxAuxData which may be attached to a transaction

Associated Types

type TxAuxData era = (r ∷ Type) | r → era

Methods

hashTxAuxData ∷ TxAuxData era → AuxiliaryDataHash (EraCrypto era)

validateTxAuxData ∷ ProtVer → TxAuxData era → Bool

Shelley

data ShelleyTxAuxData era where

Bundled Patterns

pattern ShelleyTxAuxData ∷ Era era ⇒ Map Word64 Metadatum → ShelleyTxAuxData era

data Metadatum

A generic metadatum type.

Constructors

Map ![(Metadatum, Metadatum)]

List ![Metadatum]

I !Integer

B !ByteString

S !Text

Allegra

data AllegraTxAuxData era where

Bundled Patterns

pattern AllegraTxAuxData ∷ Era era ⇒ Map Word64 Metadatum → StrictSeq (Timelock era) → AllegraTxAuxData era

Alonzo

data AlonzoTxAuxData era where

Bundled Patterns

pattern AlonzoTxAuxData ∷ Era era ⇒ Map Word64 Metadatum → StrictSeq (Timelock era) → Map Language (NonEmpty BinaryPlutus) → AuxiliaryData era

mkAlonzoTxAuxData ∷ (Foldable f, Era era) ⇒ Map Word64 Metadatum → f (AlonzoScript era) → AlonzoTxAuxData era

Helper function that will construct Auxiliary data from Metadatum map and a list of scripts.

Note that the relative order of same type scripts will be preserved.

getAlonzoTxAuxDataScripts ∷ Era era ⇒ AlonzoTxAuxData era → StrictSeq (AlonzoScript era)