Cardano.Ledger.Api.Tx.Body

Documentation

Building and inspecting transaction outputs

module Cardano.Ledger.Api.Tx.Out

Working with Timelock scripts and Plutus scripts

module Cardano.Ledger.Api.Scripts

class (EraTxOut era, EraTxCert era, EraPParams era, HashAnnotated (TxBody era) EraIndependentTxBody (EraCrypto era), DecCBOR (Annotator (TxBody era)), EncCBOR (TxBody era), ToCBOR (TxBody era), NoThunks (TxBody era), NFData (TxBody era), Show (TxBody era), Eq (TxBody era), EqRaw (TxBody era)) ⇒ EraTxBody era

Minimal complete definition

mkBasicTxBody, inputsTxBodyL, outputsTxBodyL, feeTxBodyL, withdrawalsTxBodyL, auxDataHashTxBodyL, spendableInputsTxBodyF, allInputsTxBodyF, certsTxBodyL, upgradeTxBody

Associated Types

type TxBody era = (r ∷ Type) | r → era

The body of a transaction.

mkBasicTxBody ∷ EraTxBody era ⇒ TxBody era

inputsTxBodyL ∷ EraTxBody era ⇒ Lens' (TxBody era) (Set (TxIn (EraCrypto era)))

outputsTxBodyL ∷ EraTxBody era ⇒ Lens' (TxBody era) (StrictSeq (TxOut era))

feeTxBodyL ∷ EraTxBody era ⇒ Lens' (TxBody era) Coin

withdrawalsTxBodyL ∷ EraTxBody era ⇒ Lens' (TxBody era) (Withdrawals (EraCrypto era))

newtype Withdrawals c

This is called wdrl in the spec.

Constructors

Withdrawals

Fields

unWithdrawals ∷ Map (RewardAcnt c) Coin

auxDataHashTxBodyL ∷ EraTxBody era ⇒ Lens' (TxBody era) (StrictMaybe (AuxiliaryDataHash (EraCrypto era)))

data AuxiliaryDataHash c

spendableInputsTxBodyF ∷ EraTxBody era ⇒ SimpleGetter (TxBody era) (Set (TxIn (EraCrypto era)))

This getter will produce all inputs from the UTxO map that this transaction might spend, which ones will depend on the validity of the transaction itself. Starting in Alonzo this will include collateral inputs.

allInputsTxBodyF ∷ EraTxBody era ⇒ SimpleGetter (TxBody era) (Set (TxIn (EraCrypto era)))

This getter will produce all inputs from the UTxO map that this transaction is referencing, even if some of them cannot be spent by the transaction. For example starting with Babbage era it will also include reference inputs.

evalBalanceTxBody

Arguments

∷ EraUTxO era

⇒ PParams era
Current protocol parameters

→ (Credential 'Staking (EraCrypto era) → Maybe Coin)
Lookup current deposit amount for a registered stake credential delegation. This function must produce valid answer for all of the stake credentials present in any of the DeRegKey delegation certificates in the supplied TxBody. In other words, there is no requirement to know about all of the delegation certificates in the ledger state, just the ones this transaction cares about.

→ (Credential 'DRepRole (EraCrypto era) → Maybe Coin)
Lookup current deposit amount for a registered DRep credential. This function must produce valid answer for all of the DRep credentials present in any of the UnRegDRep certificates in the supplied TxBody. In other words, there is no requirement to know about all of the DRep registrations in the ledger state, just the ones this transaction cares about.

→ (KeyHash 'StakePool (EraCrypto era) → Bool)
Check whether a pool with a supplied PoolStakeId is already registered. There is no requirement to answer this question for all stake pool credentials, just for the ones that have the registration certificates included in the supplied TxBody

→ UTxO era
The UTxO relevant to the transaction.

→ TxBody era
The transaction being evaluated for balance.

→ Value era
The difference between what the transaction consumes and what it produces.

Evaluate the difference between the value currently being consumed by a transaction and the total value being produced. This value will be zero for a valid transaction.

In case when full CertState is available then this can be simplified to:

let lookupRefund = lookupDepositDState (certDState dpState)
let isRegPoolId = (`Map.member` psStakePoolParams (certPState dpState))
evalBalanceTxBody pp lookupRefund isRegPoolId utxo txBody

txIdTxBody ∷ EraTxBody era ⇒ TxBody era → TxId (EraCrypto era)

Shelley Era

class (ShelleyEraTxCert era, EraTxBody era, ProtVerAtMost era 8) ⇒ ShelleyEraTxBody era

Minimal complete definition

ttlTxBodyL, updateTxBodyL

ttlTxBodyL ∷ ShelleyEraTxBody era ⇒ Lens' (TxBody era) SlotNo

updateTxBodyL ∷ ShelleyEraTxBody era ⇒ Lens' (TxBody era) (StrictMaybe (Update era))

certsTxBodyL ∷ EraTxBody era ⇒ Lens' (TxBody era) (StrictSeq (TxCert era))

Allegra Era

class EraTxBody era ⇒ AllegraEraTxBody era

Minimal complete definition

vldtTxBodyL

vldtTxBodyL ∷ AllegraEraTxBody era ⇒ Lens' (TxBody era) ValidityInterval

data ValidityInterval

ValidityInterval is a half open interval. Closed on the bottom, open on the top. A SNothing on the bottom is negative infinity, and a SNothing on the top is positive infinity

Constructors

ValidityInterval

Fields

invalidBefore ∷ !(StrictMaybe SlotNo)

invalidHereafter ∷ !(StrictMaybe SlotNo)

invalidBeforeL ∷ Lens' ValidityInterval (Maybe SlotNo)

Lens to access the invalidBefore field of a ValidityInterval as a 'Maybe SlotNo'.

invalidHereAfterL ∷ Lens' ValidityInterval (Maybe SlotNo)

Lens to access the invalidHereAfter field of a ValidityInterval as a 'Maybe SlotNo'.

Mary Era

class AllegraEraTxBody era ⇒ MaryEraTxBody era

Minimal complete definition

mintTxBodyL, mintValueTxBodyF, mintedTxBodyF

mintTxBodyL ∷ MaryEraTxBody era ⇒ Lens' (TxBody era) (MultiAsset (EraCrypto era))

mintValueTxBodyF ∷ MaryEraTxBody era ⇒ SimpleGetter (TxBody era) (Value era)

mintedTxBodyF ∷ MaryEraTxBody era ⇒ SimpleGetter (TxBody era) (Set (PolicyID (EraCrypto era)))

Alonzo Era

class (MaryEraTxBody era, AlonzoEraTxOut era) ⇒ AlonzoEraTxBody era

Minimal complete definition

collateralInputsTxBodyL, reqSignerHashesTxBodyL, scriptIntegrityHashTxBodyL, networkIdTxBodyL, redeemerPointer, redeemerPointerInverse

collateralInputsTxBodyL ∷ AlonzoEraTxBody era ⇒ Lens' (TxBody era) (Set (TxIn (EraCrypto era)))

reqSignerHashesTxBodyL ∷ AlonzoEraTxBody era ⇒ Lens' (TxBody era) (Set (KeyHash 'Witness (EraCrypto era)))

scriptIntegrityHashTxBodyL ∷ AlonzoEraTxBody era ⇒ Lens' (TxBody era) (StrictMaybe (ScriptIntegrityHash (EraCrypto era)))

networkIdTxBodyL ∷ AlonzoEraTxBody era ⇒ Lens' (TxBody era) (StrictMaybe Network)

redeemerPointer ∷ AlonzoEraTxBody era ⇒ TxBody era → PlutusPurpose AsItem era → StrictMaybe (PlutusPurpose AsIx era)

This function is called rdptr in the spec. Given a TxBody and a plutus purpose with an item, we should be able to find the plutus purpose as in index

redeemerPointerInverse ∷ AlonzoEraTxBody era ⇒ TxBody era → PlutusPurpose AsIx era → StrictMaybe (PlutusPurpose AsIxItem era)

This is an inverse of redeemerPointer. Given purpose as an index return it as an item.

Babbage Era

class (AlonzoEraTxBody era, BabbageEraTxOut era) ⇒ BabbageEraTxBody era

Minimal complete definition

sizedOutputsTxBodyL, referenceInputsTxBodyL, totalCollateralTxBodyL, collateralReturnTxBodyL, sizedCollateralReturnTxBodyL, allSizedOutputsTxBodyF

sizedOutputsTxBodyL ∷ BabbageEraTxBody era ⇒ Lens' (TxBody era) (StrictSeq (Sized (TxOut era)))

referenceInputsTxBodyL ∷ BabbageEraTxBody era ⇒ Lens' (TxBody era) (Set (TxIn (EraCrypto era)))

totalCollateralTxBodyL ∷ BabbageEraTxBody era ⇒ Lens' (TxBody era) (StrictMaybe Coin)

collateralReturnTxBodyL ∷ BabbageEraTxBody era ⇒ Lens' (TxBody era) (StrictMaybe (TxOut era))

sizedCollateralReturnTxBodyL ∷ BabbageEraTxBody era ⇒ Lens' (TxBody era) (StrictMaybe (Sized (TxOut era)))

allSizedOutputsTxBodyF ∷ BabbageEraTxBody era ⇒ SimpleGetter (TxBody era) (StrictSeq (Sized (TxOut era)))

Conway Era

class (BabbageEraTxBody era, ConwayEraTxCert era, ConwayEraPParams era, ConwayEraScript era) ⇒ ConwayEraTxBody era

Minimal complete definition

currentTreasuryValueTxBodyL, votingProceduresTxBodyL, proposalProceduresTxBodyL, treasuryDonationTxBodyL

votingProceduresTxBodyL ∷ ConwayEraTxBody era ⇒ Lens' (TxBody era) (VotingProcedures era)

Lens for getting and setting VotingProcedures.

proposalProceduresTxBodyL ∷ ConwayEraTxBody era ⇒ Lens' (TxBody era) (OSet (ProposalProcedure era))

Lens for getting and setting ProposalProcedures.