Cardano.Ledger.Api.Tx.Body

Documentation

Building and inspecting transaction outputs

module Cardano.Ledger.Api.Tx.Out

Working with Timelock scripts and Plutus scripts

module Cardano.Ledger.Api.Scripts

class (EraTxOut era, EraPParams era, HashAnnotated (TxBody era) EraIndependentTxBody (EraCrypto era), DecCBOR (Annotator (TxBody era)), EncCBOR (TxBody era), ToCBOR (TxBody era), NoThunks (TxBody era), NFData (TxBody era), Show (TxBody era), Eq (TxBody era)) ⇒ EraTxBody era

Minimal complete definition

mkBasicTxBody, inputsTxBodyL, outputsTxBodyL, feeTxBodyL, withdrawalsTxBodyL, auxDataHashTxBodyL, allInputsTxBodyF

Associated Types

type TxBody era = (r ∷ Type) | r → era

The body of a transaction.

mkBasicTxBody ∷ EraTxBody era ⇒ TxBody era

inputsTxBodyL ∷ EraTxBody era ⇒ Lens' (TxBody era) (Set (TxIn (EraCrypto era)))

outputsTxBodyL ∷ EraTxBody era ⇒ Lens' (TxBody era) (StrictSeq (TxOut era))

feeTxBodyL ∷ EraTxBody era ⇒ Lens' (TxBody era) Coin

withdrawalsTxBodyL ∷ EraTxBody era ⇒ Lens' (TxBody era) (Withdrawals (EraCrypto era))

newtype Withdrawals c

This is called wdrl in the spec.

Constructors

Withdrawals

Fields

unWithdrawals ∷ Map (RewardAcnt c) Coin

auxDataHashTxBodyL ∷ EraTxBody era ⇒ Lens' (TxBody era) (StrictMaybe (AuxiliaryDataHash (EraCrypto era)))

data AuxiliaryDataHash c

allInputsTxBodyF ∷ EraTxBody era ⇒ SimpleGetter (TxBody era) (Set (TxIn (EraCrypto era)))

evalBalanceTxBody

Arguments

∷ (EraUTxO era, ShelleyEraTxBody era)

⇒ PParams era
Current protocol parameters

→ (StakeCredential (EraCrypto era) → Maybe Coin)
Lookup current deposit amount for a registered stake credential delegation. This function must produce valid answer for all of the stake credentials present in any of the DeRegKey delegation certificates in the supplied TxBody. In other words, there is no requirement to know about all of the delegation certificates in the ledger state, just the ones this transaction cares about.

→ (KeyHash 'StakePool (EraCrypto era) → Bool)
Check whether a pool with a supplied PoolStakeId is already registered. There is no requirement to answer this question for all stake pool credentials, just for the ones that have the registration certificates included in the supplied TxBody

→ UTxO era
The UTxO relevant to the transaction.

→ TxBody era
The transaction being evaluated for balance.

→ Value era
The difference between what the transaction consumes and what it produces.

Evaluate the difference between the value currently being consumed by a transaction and the total value being produced. This value will be zero for a valid transaction.

In case when full CertState is available then this can be simplified to:

let lookupRefund = lookupDepositDState (certDState dpState)
let isRegPoolId = (`Map.member` psStakePoolParams (certPState dpState))
evalBalanceTxBody pp lookupRefund isRegPoolId utxo txBody

Shelley Era

class EraTxBody era ⇒ ShelleyEraTxBody era

Minimal complete definition

ttlTxBodyL, updateTxBodyL, certsTxBodyL

ttlTxBodyL ∷ ShelleyEraTxBody era ⇒ Lens' (TxBody era) SlotNo

updateTxBodyL ∷ ShelleyEraTxBody era ⇒ Lens' (TxBody era) (StrictMaybe (Update era))

updateTxBodyG ∷ ShelleyEraTxBody era ⇒ SimpleGetter (TxBody era) (StrictMaybe (Update era))

certsTxBodyL ∷ ShelleyEraTxBody era ⇒ Lens' (TxBody era) (StrictSeq (DCert (EraCrypto era)))

certsTxBodyG ∷ ShelleyEraTxBody era ⇒ SimpleGetter (TxBody era) (StrictSeq (DCert (EraCrypto era)))

Allegra Era

class ShelleyEraTxBody era ⇒ AllegraEraTxBody era

Minimal complete definition

vldtTxBodyL

vldtTxBodyL ∷ AllegraEraTxBody era ⇒ Lens' (TxBody era) ValidityInterval

data ValidityInterval

ValidityInterval is a half open interval. Closed on the bottom, open on the top. A SNothing on the bottom is negative infinity, and a SNothing on the top is positive infinity

Constructors

ValidityInterval

Fields

invalidBefore ∷ !(StrictMaybe SlotNo)

invalidHereafter ∷ !(StrictMaybe SlotNo)

Mary Era

class AllegraEraTxBody era ⇒ MaryEraTxBody era

Minimal complete definition

mintTxBodyL, mintValueTxBodyF, mintedTxBodyF

mintTxBodyL ∷ MaryEraTxBody era ⇒ Lens' (TxBody era) (MultiAsset (EraCrypto era))

mintValueTxBodyF ∷ MaryEraTxBody era ⇒ SimpleGetter (TxBody era) (Value era)

mintedTxBodyF ∷ MaryEraTxBody era ⇒ SimpleGetter (TxBody era) (Set (ScriptHash (EraCrypto era)))

Alonzo Era

class (MaryEraTxBody era, AlonzoEraTxOut era) ⇒ AlonzoEraTxBody era

Minimal complete definition

collateralInputsTxBodyL, reqSignerHashesTxBodyL, scriptIntegrityHashTxBodyL, networkIdTxBodyL

collateralInputsTxBodyL ∷ AlonzoEraTxBody era ⇒ Lens' (TxBody era) (Set (TxIn (EraCrypto era)))

reqSignerHashesTxBodyL ∷ AlonzoEraTxBody era ⇒ Lens' (TxBody era) (Set (KeyHash 'Witness (EraCrypto era)))

scriptIntegrityHashTxBodyL ∷ AlonzoEraTxBody era ⇒ Lens' (TxBody era) (StrictMaybe (ScriptIntegrityHash (EraCrypto era)))

networkIdTxBodyL ∷ AlonzoEraTxBody era ⇒ Lens' (TxBody era) (StrictMaybe Network)

Babbage Era

class (AlonzoEraTxBody era, BabbageEraTxOut era) ⇒ BabbageEraTxBody era

Minimal complete definition

sizedOutputsTxBodyL, referenceInputsTxBodyL, totalCollateralTxBodyL, collateralReturnTxBodyL, sizedCollateralReturnTxBodyL, allSizedOutputsTxBodyF

sizedOutputsTxBodyL ∷ BabbageEraTxBody era ⇒ Lens' (TxBody era) (StrictSeq (Sized (TxOut era)))

referenceInputsTxBodyL ∷ BabbageEraTxBody era ⇒ Lens' (TxBody era) (Set (TxIn (EraCrypto era)))

totalCollateralTxBodyL ∷ BabbageEraTxBody era ⇒ Lens' (TxBody era) (StrictMaybe Coin)

collateralReturnTxBodyL ∷ BabbageEraTxBody era ⇒ Lens' (TxBody era) (StrictMaybe (TxOut era))

sizedCollateralReturnTxBodyL ∷ BabbageEraTxBody era ⇒ Lens' (TxBody era) (StrictMaybe (Sized (TxOut era)))

allSizedOutputsTxBodyF ∷ BabbageEraTxBody era ⇒ SimpleGetter (TxBody era) (StrictSeq (Sized (TxOut era)))

Conway Era

class BabbageEraTxBody era ⇒ ConwayEraTxBody era

Minimal complete definition

votingProceduresTxBodyL, proposalProceduresTxBodyL, conwayCertsTxBodyL

votingProceduresTxBodyL ∷ ConwayEraTxBody era ⇒ Lens' (TxBody era) (StrictSeq (VotingProcedure era))

proposalProceduresTxBodyL ∷ ConwayEraTxBody era ⇒ Lens' (TxBody era) (StrictSeq (ProposalProcedure era))

conwayCertsTxBodyL ∷ ConwayEraTxBody era ⇒ Lens' (TxBody era) (StrictSeq (ConwayDCert (EraCrypto era)))