Cardano.Ledger.Api.Tx.Cert

Documentation

class (Era era, ToJSON (TxCert era), DecCBOR (TxCert era), EncCBOR (TxCert era), ToCBOR (TxCert era), FromCBOR (TxCert era), NoThunks (TxCert era), NFData (TxCert era), Show (TxCert era), Ord (TxCert era), Eq (TxCert era)) ⇒ EraTxCert era

Minimal complete definition

upgradeTxCert, getVKeyWitnessTxCert, getScriptWitnessTxCert, mkRegPoolTxCert, getRegPoolTxCert, mkRetirePoolTxCert, getRetirePoolTxCert, lookupRegStakeTxCert, lookupUnRegStakeTxCert, getTotalDepositsTxCerts, getTotalRefundsTxCerts

Associated Types

type TxCert era = (r ∷ Type) | r → era

type TxCertUpgradeError era

type TxCertUpgradeError era = Void

upgradeTxCert ∷ EraTxCert era ⇒ TxCert (PreviousEra era) → Either (TxCertUpgradeError era) (TxCert era)

Every era, except Shelley, must be able to upgrade a TxCert from a previous era. However, not all certificates can be upgraded, because some eras lose some of the certificates, thus return type is an Either. Eg. from Babbage to Conway: MIR and Genesis certificates were removed.

getVKeyWitnessTxCert ∷ EraTxCert era ⇒ TxCert era → Maybe (KeyHash 'Witness (EraCrypto era))

Return a witness key whenever a certificate requires one

getScriptWitnessTxCert ∷ EraTxCert era ⇒ TxCert era → Maybe (ScriptHash (EraCrypto era))

Return a ScriptHash for certificate types that require a witness

pattern RegPoolTxCert ∷ EraTxCert era ⇒ PoolParams (EraCrypto era) → TxCert era

pattern RetirePoolTxCert ∷ EraTxCert era ⇒ KeyHash 'StakePool (EraCrypto era) → EpochNo → TxCert era

lookupRegStakeTxCert ∷ EraTxCert era ⇒ TxCert era → Maybe (Credential 'Staking (EraCrypto era))

Extract staking credential from any certificate that can register such credential

lookupUnRegStakeTxCert ∷ EraTxCert era ⇒ TxCert era → Maybe (Credential 'Staking (EraCrypto era))

Extract staking credential from any certificate that can unregister such credential

isRegStakeTxCert ∷ EraTxCert era ⇒ TxCert era → Bool

Check if supplied TxCert is a stake registering certificate

isUnRegStakeTxCert ∷ EraTxCert era ⇒ TxCert era → Bool

Check if supplied TxCert is a stake un-registering certificate

Shelley Era

Complete set of patterns for Shelley through Babbage TxCert:

TxCert =
  RegPoolTxCert
  RetirePoolTxCert
  RegTxCert
  UnRegTxCert
  DelegStakeTxCert
  MirTxCert
  GenesisDelegTxCert

class EraTxCert era ⇒ ShelleyEraTxCert era

Minimal complete definition

mkRegTxCert, getRegTxCert, mkUnRegTxCert, getUnRegTxCert, mkDelegStakeTxCert, getDelegStakeTxCert, mkGenesisDelegTxCert, getGenesisDelegTxCert, mkMirTxCert, getMirTxCert

pattern MirTxCert ∷ (ShelleyEraTxCert era, ProtVerAtMost era 8) ⇒ MIRCert (EraCrypto era) → TxCert era

pattern GenesisDelegTxCert ∷ (ShelleyEraTxCert era, ProtVerAtMost era 8) ⇒ KeyHash 'Genesis (EraCrypto era) → KeyHash 'GenesisDelegate (EraCrypto era) → Hash (EraCrypto era) (VerKeyVRF (EraCrypto era)) → TxCert era

pattern RegTxCert ∷ ShelleyEraTxCert era ⇒ StakeCredential (EraCrypto era) → TxCert era

pattern UnRegTxCert ∷ ShelleyEraTxCert era ⇒ StakeCredential (EraCrypto era) → TxCert era

pattern DelegStakeTxCert ∷ ShelleyEraTxCert era ⇒ StakeCredential (EraCrypto era) → KeyHash 'StakePool (EraCrypto era) → TxCert era

Conway Era

Complete set of patterns for Conway TxCert:

TxCert =
  RegPoolTxCert
  RetirePoolTxCert
  RegTxCert
  UnRegTxCert
  RegDepositTxCert
  UnRegDepositTxCert
  DelegTxCert
  RegDepositDelegTxCert
  AuthCommitteeHotKeyTxCert
  ResignCommitteeColdTxCert
  RegDRepTxCert
  UnRegDRepTxCert

class ShelleyEraTxCert era ⇒ ConwayEraTxCert era

Minimal complete definition

mkRegDepositTxCert, getRegDepositTxCert, mkUnRegDepositTxCert, getUnRegDepositTxCert, mkDelegTxCert, getDelegTxCert, mkRegDepositDelegTxCert, getRegDepositDelegTxCert, mkAuthCommitteeHotKeyTxCert, getAuthCommitteeHotKeyTxCert, mkResignCommitteeColdTxCert, getResignCommitteeColdTxCert, mkRegDRepTxCert, getRegDRepTxCert, mkUnRegDRepTxCert, getUnRegDRepTxCert, mkUpdateDRepTxCert, getUpdateDRepTxCert

getDelegateeTxCert ∷ ConwayEraTxCert era ⇒ TxCert era → Maybe (Delegatee (EraCrypto era))

data Delegatee c

First type argument is the deposit

Constructors

DelegStake !(KeyHash 'StakePool c)

DelegVote !(DRep c)

DelegStakeVote !(KeyHash 'StakePool c) !(DRep c)

getStakePoolDelegatee ∷ Delegatee c → Maybe (KeyHash 'StakePool c)

pattern RegDepositTxCert ∷ ConwayEraTxCert era ⇒ StakeCredential (EraCrypto era) → Coin → TxCert era

pattern UnRegDepositTxCert ∷ ConwayEraTxCert era ⇒ StakeCredential (EraCrypto era) → Coin → TxCert era

pattern DelegTxCert ∷ ConwayEraTxCert era ⇒ StakeCredential (EraCrypto era) → Delegatee (EraCrypto era) → TxCert era

pattern RegDepositDelegTxCert ∷ ConwayEraTxCert era ⇒ StakeCredential (EraCrypto era) → Delegatee (EraCrypto era) → Coin → TxCert era

pattern AuthCommitteeHotKeyTxCert ∷ ConwayEraTxCert era ⇒ Credential 'ColdCommitteeRole (EraCrypto era) → Credential 'HotCommitteeRole (EraCrypto era) → TxCert era

pattern ResignCommitteeColdTxCert ∷ ConwayEraTxCert era ⇒ Credential 'ColdCommitteeRole (EraCrypto era) → StrictMaybe (Anchor (EraCrypto era)) → TxCert era

pattern RegDRepTxCert ∷ ConwayEraTxCert era ⇒ Credential 'DRepRole (EraCrypto era) → Coin → StrictMaybe (Anchor (EraCrypto era)) → TxCert era

pattern UnRegDRepTxCert ∷ ConwayEraTxCert era ⇒ Credential 'DRepRole (EraCrypto era) → Coin → TxCert era