Cardano.Ledger.Api.Tx.Out

Description

This module is used for building and inspecting transaction outputs.

You'll find some examples below.

Let's start by defining the GHC extensions and imports.

>>> :set -XTypeApplications
>>> import Cardano.Ledger.Api.Era (Babbage)
>>> import Lens.Micro
>>> import Test.Cardano.Ledger.Babbage.Serialisation.Generators () -- Neded for doctests only
>>> import Test.QuickCheck -- Needed for doctests only

Here's an example on how to build a very basic Babbage era transaction output with a random address and value, and without any datum or reference script.

>>> :{
quickCheck $ \addr val ->
    let
        -- Defining a Babbage era transaction output with some random address and value.
        txOut = mkBasicTxOut @Babbage addr val
     in
        -- We verify that the transaction output contains our random address and value.
        txOut ^. addrTxOutL == addr && txOut ^. valueTxOutL == val
:}
+++ OK, passed 100 tests.

Documentation

module Cardano.Ledger.Api.Tx.Address

class (Val (Value era), ToJSON (TxOut era), DecCBOR (Value era), DecCBOR (CompactForm (Value era)), EncCBOR (Value era), ToCBOR (TxOut era), FromCBOR (TxOut era), EncCBOR (TxOut era), DecCBOR (TxOut era), DecShareCBOR (TxOut era), Share (TxOut era) ~ Interns (Credential 'Staking (EraCrypto era)), NoThunks (TxOut era), NFData (TxOut era), Show (TxOut era), Eq (TxOut era), EraPParams era) ⇒ EraTxOut era

Abstract interface into specific fields of a TxOut

Minimal complete definition

mkBasicTxOut, upgradeTxOut, valueEitherTxOutL, addrEitherTxOutL, (getMinCoinSizedTxOut | getMinCoinTxOut)

Associated Types

type TxOut era = (r ∷ Type) | r → era

The output of a UTxO for a particular era

mkBasicTxOut ∷ EraTxOut era ⇒ Addr (EraCrypto era) → Value era → TxOut era

upgradeTxOut ∷ EraTxOut era ⇒ TxOut (PreviousEra era) → TxOut era

Every era, except Shelley, must be able to upgrade a TxOut from a previous era.

Value

valueTxOutL ∷ EraTxOut era ⇒ Lens' (TxOut era) (Value era)

coinTxOutL ∷ (HasCallStack, EraTxOut era) ⇒ Lens' (TxOut era) Coin

isAdaOnlyTxOutF ∷ EraTxOut era ⇒ SimpleGetter (TxOut era) Bool

This is a getter that implements an efficient way to check whether TxOut contains ADA only.

Address

addrTxOutL ∷ EraTxOut era ⇒ Lens' (TxOut era) (Addr (EraCrypto era))

bootAddrTxOutF ∷ EraTxOut era ⇒ SimpleGetter (TxOut era) (Maybe (BootstrapAddress (EraCrypto era)))

Size

getMinCoinTxOut ∷ EraTxOut era ⇒ PParams era → TxOut era → Coin

Same as getMinCoinSizedTxOut, except information about the size of TxOut will be computed by serializing the TxOut. If the size turns out to be not needed, then serialization will have no overhead, since it is computed lazily.

setMinCoinTxOut ∷ EraTxOut era ⇒ PParams era → TxOut era → TxOut era

Same as setMinCoinSizedTxOut, except it doesn't require the size of the TxOut and will recompute it if needed. Initial amount is not important.

getMinCoinSizedTxOut ∷ EraTxOut era ⇒ PParams era → Sized (TxOut era) → Coin

Produce the minimum lovelace that a given transaction output must contain. Information about the size of the TxOut is required in some eras. Use getMinCoinTxOut if you don't have the size readily available to you.

setMinCoinSizedTxOut ∷ ∀ era. EraTxOut era ⇒ PParams era → Sized (TxOut era) → Sized (TxOut era)

This function will adjust the output's Coin value to the smallest amount allowed by the UTXO rule. Initial amount is not important.

ensureMinCoinTxOut ∷ EraTxOut era ⇒ PParams era → TxOut era → TxOut era

Similar to setMinCoinTxOut it will guarantee that the minimum requirement for the output amount is satisified, however it makes it possible to set a higher amount than the minimaly required.

> ensureMinCoinTxOut pp (txOut & coinTxOutL .~ zero) == setMinCoinTxOut pp (txOut & coinTxOutL .~ zero)
> (ensureMinCoinTxOut pp txOut ^. coinTxOutL) >= (setMinCoinTxOut pp txOut ^. coinTxOutL)

ensureMinCoinSizedTxOut ∷ ∀ era. EraTxOut era ⇒ PParams era → Sized (TxOut era) → Sized (TxOut era)

Similar to setMinCoinSizedTxOut it will guarantee that the minimum requirement for the output amount is satisified, however it makes it possible to set a higher amount than the minimaly required.

ensureMinCoinSizedTxOut relates to setMinCoinSizedTxOut in the same way that ensureMinCoinTxOut relates to setMinCoinTxOut.

Shelley, Allegra and Mary Era

Alonzo Era

class (AlonzoEraPParams era, EraTxOut era) ⇒ AlonzoEraTxOut era

Minimal complete definition

dataHashTxOutL, datumTxOutF

dataHashTxOutL ∷ AlonzoEraTxOut era ⇒ Lens' (TxOut era) (StrictMaybe (DataHash (EraCrypto era)))

type DataHash c = SafeHash c EraIndependentData

datumTxOutF ∷ AlonzoEraTxOut era ⇒ SimpleGetter (TxOut era) (Datum era)

Babbage Era

class (AlonzoEraTxOut era, AlonzoEraScript era) ⇒ BabbageEraTxOut era

Minimal complete definition

referenceScriptTxOutL, dataTxOutL, datumTxOutL

dataTxOutL ∷ BabbageEraTxOut era ⇒ Lens' (TxOut era) (StrictMaybe (Data era))

data Data era where

Bundled Patterns

pattern Data ∷ Era era ⇒ Data → Data era

datumTxOutL ∷ BabbageEraTxOut era ⇒ Lens' (TxOut era) (Datum era)

data Datum era

Datum can be described by a either a data hash or binary data, but not both. It can also be neither one of them.

Constructors

NoDatum

DatumHash !(DataHash (EraCrypto era))

Datum !(BinaryData era)

referenceScriptTxOutL ∷ BabbageEraTxOut era ⇒ Lens' (TxOut era) (StrictMaybe (Script era))