Cardano.Ledger.Api.Tx.Wits

Shelley onwards

class (EraScript era, Eq (TxWits era), EqRaw (TxWits era), Show (TxWits era), Monoid (TxWits era), NoThunks (TxWits era), ToCBOR (TxWits era), EncCBOR (TxWits era), DecCBOR (Annotator (TxWits era))) ⇒ EraTxWits era

A collection of witnesses in a Tx

Minimal complete definition

addrTxWitsL, bootAddrTxWitsL, scriptTxWitsL, upgradeTxWits

Associated Types

type TxWits era = (r ∷ Type) | r → era

mkBasicTxWits ∷ EraTxWits era ⇒ TxWits era

Address witness

addrTxWitsL ∷ EraTxWits era ⇒ Lens' (TxWits era) (Set (WitVKey 'Witness (EraCrypto era)))

data KeyRole

The role of a key.

Note that a role is not _fixed_, nor is it unique. In particular, keys may variously be used as witnesses, and so in many case we will change the role of a key to the Witness role.

It is also perfectly allowable for a key to be used in many roles; there is nothing prohibiting somebody using the same underlying key as their payment and staking key, as well as the key for their stake pool. So these roles are more intended for two purposes:

To make explicit how we are using a key in the specifications

To provide a guide to downstream implementors, for whom the profusion of keys may be confusing.

Constructors

Witness

WitVKey

data WitVKey (kr ∷ KeyRole) c where

Proof/Witness that a transaction is authorized by the given key holder.

Bundled Patterns

pattern WitVKey ∷ (Typeable kr, Crypto c) ⇒ VKey kr c → SignedDSIGN c (Hash c EraIndependentTxBody) → WitVKey kr c

witVKeyBytes ∷ ∀ (kr ∷ KeyRole) c. WitVKey kr c → ByteString

Access CBOR encoded representation of the witness. Evaluated lazily

witVKeyHash ∷ ∀ (kr ∷ KeyRole) c. WitVKey kr c → KeyHash 'Witness c

Access computed hash. Evaluated lazily

Byron address witness

bootAddrTxWitsL ∷ EraTxWits era ⇒ Lens' (TxWits era) (Set (BootstrapWitness (EraCrypto era)))

data BootstrapWitness c

Script witness

scriptTxWitsL ∷ EraTxWits era ⇒ Lens' (TxWits era) (Map (ScriptHash (EraCrypto era)) (Script era))

hashScriptTxWitsL ∷ EraTxWits era ⇒ Lens (TxWits era) (TxWits era) (Map (ScriptHash (EraCrypto era)) (Script era)) [Script era]

This is a helper lens that will hash the scripts when adding as witnesses.

Alonzo onwards

class (EraTxWits era, AlonzoEraScript era) ⇒ AlonzoEraTxWits era

Minimal complete definition

datsTxWitsL, rdmrsTxWitsL

datsTxWitsL ∷ AlonzoEraTxWits era ⇒ Lens' (TxWits era) (TxDats era)

hashDataTxWitsL ∷ AlonzoEraTxWits era ⇒ Lens (TxWits era) (TxWits era) (TxDats era) [Data era]

This is a convenience Lens that will hash the Data when it is being added to the TxWits. See datsTxWitsL for a version that aloows setting TxDats instead.

data TxDats era where

Note that TxDats are based on MemoBytes since we must preserve the original bytes for the ScriptIntegrity. Since the TxDats exist outside of the transaction body, this is how we ensure that they are not manipulated.

Bundled Patterns

pattern TxDats ∷ Era era ⇒ Map (DataHash (EraCrypto era)) (Data era) → TxDats era

pattern TxDats' ∷ Map (DataHash (EraCrypto era)) (Data era) → TxDats era

unTxDats ∷ TxDats era → Map (DataHash (EraCrypto era)) (Data era)

rdmrsTxWitsL ∷ AlonzoEraTxWits era ⇒ Lens' (TxWits era) (Redeemers era)

data Redeemers era where

Note that Redeemers are based on MemoBytes since we must preserve the original bytes for the ScriptIntegrity. Since the Redeemers exist outside of the transaction body, this is how we ensure that they are not manipulated.

Bundled Patterns

pattern Redeemers ∷ AlonzoEraScript era ⇒ Map (PlutusPurpose AsIx era) (Data era, ExUnits) → Redeemers era

unRedeemers ∷ Redeemers era → Map (PlutusPurpose AsIx era) (Data era, ExUnits)

type family PlutusPurpose (f ∷ Type → Type → Type) era = (r ∷ Type) | r → era

data AlonzoPlutusPurpose (f ∷ Type → Type → Type) era

Constructors

AlonzoSpending !(f Word32 (TxIn (EraCrypto era)))

AlonzoMinting !(f Word32 (PolicyID (EraCrypto era)))

AlonzoCertifying !(f Word32 (TxCert era))

AlonzoRewarding !(f Word32 (RewardAccount (EraCrypto era)))

newtype AsIx ix it

Constructors

AsIx

Fields

unAsIx ∷ ix

newtype AsItem ix it

Constructors

AsItem

Fields

unAsItem ∷ it

Conway

data ConwayPlutusPurpose (f ∷ Type → Type → Type) era

Constructors

ConwaySpending !(f Word32 (TxIn (EraCrypto era)))

ConwayMinting !(f Word32 (PolicyID (EraCrypto era)))

ConwayCertifying !(f Word32 (TxCert era))

ConwayRewarding !(f Word32 (RewardAccount (EraCrypto era)))

ConwayVoting !(f Word32 (Voter (EraCrypto era)))

ConwayProposing !(f Word32 (ProposalProcedure era))