Cardano.Ledger.Api.UTxO

Documentation

newtype UTxO era

The unspent transaction outputs.

Constructors

UTxO

Fields

unUTxO ∷ Map (TxIn (EraCrypto era)) (TxOut era)

class EraTxBody era ⇒ EraUTxO era where

Associated Types

type ScriptsNeeded era = (r ∷ Type) | r → era

A customizable type on per era basis for the information required to find all scripts needed for the transaction.

Methods

getConsumedValue

Arguments

∷ PParams era

→ (StakeCredential (EraCrypto era) → Maybe Coin)
Function that can lookup current delegation deposits

→ UTxO era

→ TxBody era

→ Value era

Calculate all the value that is being consumed by the transaction.

getScriptsNeeded ∷ UTxO era → TxBody era → ScriptsNeeded era

Produce all the information required for figuring out which scripts are required for the transaction to be valid, once those scripts are evaluated

getScriptsHashesNeeded ∷ ScriptsNeeded era → Set (ScriptHash (EraCrypto era))

Extract the set of all script hashes that are needed for script validation.